% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversities.R
\name{abundancePerSample}
\alias{abundancePerSample}
\title{abundance (reads, gamme0) per sample
return 1-column data frame}
\usage{
abundancePerSample(t.community.matrix, hasTotal = TRUE)
}
\description{
abundance (reads, gamme0) per sample
return 1-column data frame
}

