% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethylHelpers.R
\name{addSampleToExperiment}
\alias{addSampleToExperiment}
\title{Add a microbeMethyl object to a microbeMethylExperiment object}
\usage{
addSampleToExperiment(experiment, sample)
}
\arguments{
\item{experiment}{A microbeMethylExperiment object.}

\item{sample}{A microbeMethyl object.}
}
\value{
The updated microbeMethylExperiment object.
}
\description{
Adds a microbeMethyl object to the list of samples in a
microbeMethylExperiment object.
}
