% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{annotation,commaData-method}
\alias{annotation,commaData-method}
\title{Accessor for genomic feature annotation}
\usage{
\S4method{annotation}{commaData}(object)
}
\arguments{
\item{object}{A \code{commaData} object.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object. May be empty (length
0) if no annotation was provided when creating the object.
}
\description{
Returns the \code{\link[GenomicRanges]{GRanges}} of genomic features stored
in a \code{\link{commaData}} object. This is the annotation loaded from a
GFF3 or BED file at construction time.
}
\examples{
data(comma_example_data)
annotation(comma_example_data)

}
