% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{as.type}
\alias{as.type}
\title{Convert data frame columns to different type, 
default to numeric type for columns.}
\source{
Modified from 
\url{http://stackoverflow.com/questions/2288485/how-to-convert-a-data-frame-column-to-numeric-type}.
}
\usage{
as.type(df, FUN = as.numeric)
}
\arguments{
\item{df}{A data frame.}

\item{MARGIN, FUN}{Refer to \code{\link{apply}}.}
}
\description{
Convert data frame columns to different type, 
default to numeric type for columns.
}
\examples{
df <- as.type(df)
df <- as.type(df, FUN=as.character)
}
\keyword{utils}

