% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMethylRawList.R
\name{buildMethylRawList}
\alias{buildMethylRawList}
\title{Build methylRawList object}
\usage{
buildMethylRawList(
  df,
  assembly = "NC_000913.3",
  context = "GATC",
  resolution = "base"
)
}
\arguments{
\item{df}{data frame containing columns starting with "cov_" and "methyl_"
for each sample (i.e. "cov_sample1")}

\item{assembly}{a string, chromosome assembly, only relevant if there are
multiple chroms otherwise any string}

\item{context}{a string, sequence context around methylated base}

\item{resolution}{a string, either "base" or "region"}
}
\value{
a MethylRawList object ({methylKit} S4 class)
}
\description{
Build methylRawList object
}
