% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityAnalysis.R
\name{classification}
\alias{classification}
\alias{ctreeClassification}
\alias{lassoClassification}
\title{Classify samples by taxonomic composition}
\usage{
ctreeClassification(comm, attr.data, group.id, levels = c(), ...)

lassoClassification(relative.abund, attr.data, group.id, percent = 0.01,
  alpha = 1, family = "multinomial", nlambda = 500,
  coef.s = "lambda.min", return.df = TRUE, ...)
}
\arguments{
\item{comm}{A community matrix, which can be either abundance or relative abundance.}

\item{group.id}{The column name in \code{attr.data} contains the known groups to 
compare with enterotypes.}

\item{levels}{Levels to order data by \code{group.id}.}

\item{relative.abund, attr.data, percent}{Refer to \code{\link{enterotype}}.}

\item{alpha, family, nlambda, ...}{Refer to \code{\link{cv.glmnet}}.}

\item{coef.s}{Which coeffient to return by \code{\link{coef}}.
\code{lambda.min}, as default, is the value of lambda that 
gives minimum mean cross-validated error. 
\code{lambda.1se} gives the most regularized model such that 
error is within one standard error of the minimum.}
}
\description{
Build a classification model to classify or predict 
the certain type of samples (e.g. land types).
}
\details{
Create a conditional inference tree \code{\link{ctree}} for the classification of samples.
Ctree uses a significance test procedure in order to select variables 
instead of selecting the variable that maximizes an information measure (e.g. Gini coefficient).
\url{http://stats.stackexchange.com/questions/12140/conditional-inference-trees-vs-traditional-decision-trees}.

Build a classification model using lasso \code{\link{cv.glmnet}},
given relative abundance of taxonomic compositions (e.g. families) in the samples, 
and also to find which taxonomic compositions are important.
}
\examples{
model <- ctreeClassification(comm, attr.data=env, group.id="land.use")
plot(model$ctree)

cvfit <- lassoClassification(relative.abund, attr.data=env, group.id="land.use")

}
\keyword{classification}

