% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityAnalysis.R
\name{classifyByTaxaComp}
\alias{classifyByTaxaComp}
\title{Classify sample attribute by taxonomic composition}
\usage{
classifyByTaxaComp(data, attr.data, group.id, percent = 0.01, alpha = 1,
  family = "multinomial", nlambda = 500, coef.s = "lambda.min",
  return.df = TRUE, ...)
}
\arguments{
\item{data, attr.data, group.id, percent}{Refer to \code{\link{enterotype}}.}

\item{alpha, family, nlambda, ...}{Refer to \code{\link{cv.glmnet}}.}

\item{coef.s}{Which coeffient to return by \code{\link{coef}}.
\code{lambda.min}, as default, is the value of lambda that 
gives minimum mean cross-validated error. 
\code{lambda.1se} gives the most regularized model such that 
error is within one standard error of the minimum.}
}
\details{
Build a classification model using lasso \code{\link{cv.glmnet}},
to predict the certain type of samples (e.g. land types) 
given relative abundance of taxonomic compositions (e.g. families) in the samples, 
and also to find which taxonomic compositions are important to express this model.

This can be used after
}
\examples{
cvfit <- classifyByTaxaComp(abun.dist.matrix, attr.data=env, group.id="land.use")
}
\keyword{classification}

