% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMStatistics.R
\name{CMStatistics}
\alias{CMStatistics}
\alias{summaryCM}
\alias{summaryCM.Vector}
\alias{summaryDiversity}
\alias{summaryOTUs}
\alias{summaryTaxaGroup}
\title{Statistics to one or multipe community matrix}
\usage{
summaryCM.Vector(community.matrix)

summaryCM(community.matrix, most.abund, has.total = 1, digits = 2,
  pretty.numbers = TRUE, x.lab = "sample", y.lab = "OTU",
  abundance.lab = "read")

summaryOTUs(..., digits = 2, input.list = FALSE, pretty.numbers = TRUE,
  x.lab = "sample", y.lab = "OTU", abundance.lab = "read")

summaryDiversity(..., row.order = c(), digits = 2, input.list = FALSE,
  pretty.numbers = TRUE, verbose = TRUE, row.names = c("$^0D_\\\\gamma$",
  "$^0D_\\\\alpha$", "$^0D_\\\\beta$", "$^1D_\\\\gamma$", "$^1D_\\\\alpha$",
  "$^1D_\\\\beta$", "$^2D_\\\\gamma$", "$^2D_\\\\alpha$", "$^2D_\\\\beta$"))

summaryTaxaGroup(..., input.list = FALSE, unclassified = 3,
  pretty.numbers = TRUE, taxa.group = c("ARCHAEA", "BACTERIA", "CHROMISTA",
  "PROTOZOA", "FUNGI", "PLANTAE", "ANIMALIA"), group.rank = "kingdom",
  count.rank = "phylum", digits = 0)
}
\arguments{
\item{most.abund}{The threshold to define the number of the most abundent OTUs.}

\item{has.total}{If 0, then only return abudence by samples (columns) of community matrix. 
If 1, then only return toal abudence. If 2, then return abudence by samples (columns) and total. 
Default to 1.}

\item{digits}{The digits to \code{\link{round}} decimal places 
if number is not interger. Default to 2.}

\item{pretty.numbers}{Default to TRUE to make numbers look pretty, 
but they will be hard to convert to numeric type.}

\item{x.lab, y.lab, abundance.lab}{The default text for "sample", "OTU", and "read".}

\item{input.list}{Default to TRUE to unwrap list(...) to 
get the actual list if the input is a list of cm.}

\item{row.order}{The same row indices of \code{row names}, 
but by a given order. Default to c().}

\item{row.names}{The row names in the summary.
Default to "reads", "OTUs", "samples", "singletons", "doubletons",
"max.OTU.abun","min.OTU.abun","max.sample.abun","min.sample.abun".}

\item{unclassified}{Refere to \code{\link{assignTaxaByRank}}, 
default to 3 to remove every rows containing "unclassified".}

\item{taxa.group}{The row names in the summary.
Default to "ARCHAEA", "BACTERIA", "CHROMISTA", "PROTOZOA", 
"FUNGI", "PLANTAE", "ANIMALIA".}

\item{group.rank}{The rank of given taxa groups, 
which determines the column name (rank) to create the subset.}

\item{count.rank}{The lower rank to be counted for each taxa group. 
Set NA to ignore this count.}

\item{row.names}{The row names in the summary.
Default to "$^0D_\\gamma$","$^0D_\\alpha$","$^0D_\\beta$",
"$^1D_\\gamma$","$^1D_\\alpha$", "$^1D_\\beta$",
"$^2D_\\gamma$","$^2D_\\alpha$","$^2D_\\beta$".}
}
\description{
Statistics to one or multipe community matrix, 
such as number of reads, OTUs, etc.  
Jost diversity is also included in the summary.
}
\details{
\code{summaryCM.Vector} return a named vector of summary of the 
community matrix, where \code{community.matrix} can be one column only.
The vector is c("reads","OTUs","samples","Shannon","singletons","doubletons").

\code{summaryCM} summarizes only single community matrix.

\code{summaryOTUs} returns a data frame of 
OTU clustering summary given one or multiple community matrix(matrices).
The summary is created by \code{\link{summaryCM.Vector}}.

\code{summaryDiversity} returns a data frame of 
OTU clustering summary given a list of community matrix.
The community matrix is transposed to an input of 
\code{\link{d}} {vegetarian}, and the summary is 
created by \code{\link{diversityTable}}.

\code{summaryTaxaGroup} returns two data frames of 
taxonomic composition of given one or multiple merged 
community matrix(matrices) with taxa table, 
which is created by \code{\link{mergeCMTaxa}}.
The 1st data frame \code{otus} summarizes the OTUs 
assigned to each taxa group.
The 2nd data frame \code{rank.count} summarizes the \code{count.rank} 
assigned to each taxa group.

The \code{group.rank} determines which taxonomic rank is based on summary.
It has to be consistent with given \code{taxa.group}. But there is a trick, 
if you want to count the number of OTUs that are identified as EUKARYOTA 
but aren't identified to Kingdom level, for example, 
you can set group.rank="kingdom" to get that number.
}
\examples{
summary.cm.vector <- summaryCM.Vector(community.matrix)

summary.cm <- summaryCM(community.matrix)

otu.stats <- summaryOTUs(cm)
# cm.list is a list of cm
otu.stats <- ComMA::summaryOTUs(cm.list, input.list=T)

div.stats <- summaryDiversity(cm, row.order=c(2,5,8,3,6,9,1,4,7))

ta.gr.stats <- summaryTaxaGroup(cm.taxa)
ta.gr.stats$otus
ta.gr.stats$rank.count 
# OTUs that were only identified to high-level EUKARYOTA ranks
ta.gr.stats <- summaryTaxaGroup(cm.taxa, group.rank="kingdom")

}
\keyword{community}
\keyword{matrix}

