% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethylHelpers.R
\name{combineMicrobeMethyls}
\alias{combineMicrobeMethyls}
\title{Combine multiple microbeMethyl objects into a microbeMethylExperiment object}
\usage{
combineMicrobeMethyls(microbeMethyl_list, experiment_metadata = list())
}
\arguments{
\item{microbeMethyl_list}{A list of microbeMethyl objects.}

\item{experiment_metadata}{A list containing metadata for the experiment (optional).}
}
\value{
A microbeMethylExperiment object containing all the input microbeMethyl objects.
}
\description{
Takes a list of microbeMethyl objects and combines them into a single
microbeMethylExperiment object.
}
