% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{combineTwoDF}
\alias{combineTwoDF}
\title{Combine two data frames or matrices with a same structure in one, 
put all values in 2nd data frame into brackets.}
\usage{
combineTwoDF(dfm, dfm2, rm.zero = TRUE, return.df = TRUE, ...)
}
\arguments{
\item{dfm}{A data frame or matrix.}

\item{dfm2}{The 2nd data frame or matrix whose values are into brackets.}

\item{rm.zero}{Default to TRUE to remove all " (0)".}

\item{return.df, ...}{Default to TRUE to return a data frame, otherwise a matrix.}
}
\description{
Combine two data frames or matrices with a same structure in one, 
put all values in 2nd data frame into brackets.
}
\examples{
df <- combineTwoDF(df, df2, stringsAsFactors=FALSE)
}
\keyword{utils}

