% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{convertType}
\alias{convertType}
\title{Convert data frame columns to different type, 
default to numeric type for columns. 
But it assign NA to actual string.}
\source{
Modified from 
\url{http://stackoverflow.com/questions/2288485/how-to-convert-a-data-frame-column-to-numeric-type}.
}
\usage{
convertType(df, FUN = as.numeric, stringsAsFactors = FALSE,
  check.names = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{MARGIN, FUN}{Refer to \code{\link{apply}}.}
}
\description{
Convert data frame columns to different type, 
default to numeric type for columns. 
But it assign NA to actual string.
}
\examples{
df <- convertType(df)
df <- convertType(df, FUN=as.character)
}
\keyword{utils}

