% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrendCoverage.R
\name{detrendCoverage}
\alias{detrendCoverage}
\title{Title Detrend Coverage of Methylation Sites by Position}
\usage{
detrendCoverage(coverage, position)
}
\arguments{
\item{coverage}{a numeric vector of coverage values for each site.}

\item{position}{a numeric vector of positions of each GATC site.}
}
\value{
a numeric vector the same length as coverage of altered coverage
values.
}
\description{
Title Detrend Coverage of Methylation Sites by Position
}
\examples{
cov <- detrendCoverage(WT_average$cov, WT_average$Position)
}
