% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{distance}
\alias{TurnoverDist}
\alias{beta1minus1}
\alias{distance}
\alias{getDissimilarity}
\title{Calculate various multivariate distance metrics}
\usage{
getDissimilarity(cm, tree = NA, is.transposed = FALSE, progressBar = TRUE,
  method = c("beta1-1", "jaccard", "horn.morisita", "bray.curtis",
  "unwt.unif", "wt.unif"))

beta1minus1(cm, is.transposed = FALSE, progressBar = TRUE,
  return.dist = TRUE)

TurnoverDist(t.community.matrix)
}
\arguments{
\item{cm}{The community matrix.}

\item{tree}{The \pkg{\link{ape}} tree object required 
to caculate \code{\link{UniFrac}}. It cannot be NA, as default, 
for method="unwt.unif" or "wt.unif".}

\item{is.transposed}{If TRUE, then the community matrix is already
transposed to be the valid input of \code{\link{vegdist}}.  
Default to FASLE.}

\item{progressBar}{Whether to print progress bar 
only when method="beta1-1" calculated by \code{\link{beta1minus1}}, 
default to TRUE.}

\item{method}{The method to calculate similarity/dissimilarity distance. 
The options are jaccard, horn.morisita, 
bray.curtis, beta1-1, and unwt.unif, wt.unif. 
The last two are unweighted or weighted \code{\link{UniFrac}} 
from \pkg{\link{phyloseq}}. 
Default to beta1-1, but it is slow.}

\item{return.dist}{If TRUE, as default, 
return a \code{\link{dist}} object, 
otherwise return a \code{\link{matrix}}.}
}
\value{
\code{getDissimilarity} returns a \code{\link{dist}} object of 
similarity/dissimilarity, whose rows and columns are 
sample names from the community matrix.
}
\description{
Data input \strong{cm} is community matrix 
defined in \pkg{\link{ComMA}}.
Rows are OTUs or individual species, and columns are sites or samples.  
It is a transposed matrix of \code{\link{vegdist}} input.
}
\details{
\code{getDissimilarity} calculates various 
multivariate distance metrics among samples.

\code{beta1minus1} calculates beta1-1.

\code{TurnoverDist} returns a \code{\link{dist}} composed of pair-wise turnovers.
}
\examples{
dist.jaccard <- getDissimilarity(cm, method="jaccard")

turnover.dist <- beta1minus1(cm)

turnover.dist <- TurnoverDist(t.community.matrix)

}
\keyword{distance}

