% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotation.R
\name{.annotationFileExt}
\alias{.annotationFileExt}
\title{Detect annotation file format from file extension}
\usage{
.annotationFileExt(file)
}
\value{
Character string, either \code{"gff"} or \code{"bed"}.
}
\description{
Detect annotation file format from file extension
}
\keyword{internal}
