% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotation.R
\name{.standardizeAnnotationMcols}
\alias{.standardizeAnnotationMcols}
\title{Standardize annotation GRanges mcols to always have feature_type and name}
\usage{
.standardizeAnnotationMcols(gr, ext)
}
\value{
A \code{GRanges} object with standardized \code{feature_type} and
\code{name} metadata columns.
}
\description{
Standardize annotation GRanges mcols to always have feature_type and name
}
\keyword{internal}
