% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enterotypes.R
\name{enterotypes}
\alias{enterotypes}
\alias{getAbundanceDist}
\alias{getClusters}
\alias{getDataCluster}
\alias{getJSD}
\alias{plotEnterotypes}
\title{Enterotypes}
\usage{
enterotypes(taxa.assign, n.max = 20, fig.path = NULL, percent = 0.01,
  verbose = TRUE, ...)

getAbundanceDist(taxa.assign)

getJSD(abun.dist.matrix)

getClusters(abun.dist.matrix, jsd.dist, n.max = 20, fig.path = NULL, ...)

getDataCluster(jsd.dist, k = NULL, measurements = NULL, validate = TRUE,
  verbose = TRUE)

plotEnterotypes(data, data.dist, data.cluster, fig.path, percent = 0,
  addLabel = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{taxa.assign}{The data frame of taxonomic assignments with abudence
at the \code{rank}, where rownames are taxonomy at that rank, 
and columns are the sample names. It can be 
one element of the list generated by \code{\link{assignTaxaByRank}}.}

\item{n.max}{The number of clusters. Default to 20.}

\item{fig.path}{The folder path to save figures. If NULL, 
the default, do not plot the figure.}

\item{percent}{The percentage threshold to remove the noise 
(ie. low abundant genera). Prior to the analysis. 
Default to 0.}

\item{abun.dist.matrix}{The matrix of normalized probability distributions 
of the abundance matrix, also called abundance distributions.}

\item{jsd.dist}{A distance \code{\link{dist}} object 
calculated by Jensen-Shannon divergence (JSD) metric.}

\item{k}{The number of clusters chosen for the optimised solution.
If NULL, the defult, it will take the number of clusters having 
the largest CH Index.}

\item{measurements}{The vector of measurements, such as 
Calinski-Harabasz (CH) Index, to find the optimised solution.}

\item{validate}{Logical if it needs to validate. Default to TRUE.}

\item{addLabel}{Logical if add labels of points. Default to TRUE.}

\item{abun.dist.matrix}{The \code{\link{dist}} object of 
distance Jensen-Shannon divergence (JSD).}
}
\description{
The pipeline built for enterotypes 
\url{http://enterotype.embl.de}.
}
\details{
\code{enterotypes} is the pipeline summarised from 
\url{http://enterotype.embl.de/enterotypes_tutorial.sanger.R}.
The list of other functions below are the components to form 
this pipeline.

\code{getJSD} return a distance \code{\link{dist}} object 
calculated by Jensen-Shannon divergence (JSD) metric.

\code{getClusters} return the Calinski-Harabasz (CH) Index 
for choosig a number of clusters from 2 to \code{n.max}.

\code{getDataCluster} return the optimised or selected solution having 
\code{k} clusters.

\code{plotEnterotypes} is a mixed function to plot clusters, 
BCA \code{\link{bca}} and PCoA \code{\link{dudi.pco}}
from the optimised or selected solution.

Between-class analysis (BCA) was performed to support the 
clustering and identify the drivers for the enterotypes.
It is only available when \code{k} > 2.
}
\examples{
enterotypes(taxa.assign, n.max=20, fig.path="figures", percent=0)

abun.dist.matrix <- getAbundanceDist(taxa.assign)

jsd.dist <- getJSD(abun.dist.matrix)

nclusters <- getClusters(abun.dist.matrix, jsd.dist)

data.cluster <- getDataCluster(jsd.dist, measurements=nclusters)

data.cluster <- getClusters(jsd.dist)

}
\keyword{enterotypes}

