% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{genome,commaData-method}
\alias{genome,commaData-method}
\title{Accessor for genome size information}
\usage{
\S4method{genome}{commaData}(x)
}
\arguments{
\item{x}{A \code{commaData} object.}
}
\value{
A named integer vector of chromosome sizes
(chromosome name -> length in bp), or \code{NULL} if no genome information
was provided at construction.
}
\description{
Returns the chromosome sizes stored in a \code{\link{commaData}} object.
Genome size information is stored in the \code{Seqinfo} attached to
\code{rowRanges(object)}. This accessor returns \code{seqlengths(object)}
for backward compatibility.
}
\examples{
data(comma_example_data)
genome(comma_example_data)

}
