% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{getBetaValues,MicrobeMethyl-method}
\alias{getBetaValues,MicrobeMethyl-method}
\title{getBetaValues for MicrobeMethyl objects}
\usage{
\S4method{getBetaValues}{MicrobeMethyl}(object)
}
\arguments{
\item{object}{A MicrobeMethyl object}
}
\value{
A numeric vector of beta values for the given MicrobeMethyl object
}
\description{
Retrieve the beta values from a MicrobeMethyl object.
}
