% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl_classes.R
\name{getBetaValues,MicrobeMethylExperiment-method}
\alias{getBetaValues,MicrobeMethylExperiment-method}
\title{getBetaValues for MicrobeMethylExperiment objects}
\usage{
\S4method{getBetaValues}{MicrobeMethylExperiment}(object)
}
\arguments{
\item{object}{A MicrobeMethylExperiment object.}
}
\value{
A list of numeric vectors, where each vector contains the percent
methylation values for a sample in the MicrobeMethylExperiment object.
}
\description{
Retrieve the percent methylation values for all samples in a
MicrobeMethylExperiment object.
}
