% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl_classes.R
\name{getBetaValues}
\alias{getBetaValues}
\title{Get percent methylation values from MicrobeMethyl or MicrobeMethylExperiment}
\usage{
getBetaValues(object)
}
\arguments{
\item{object}{An object of class MicrobeMethyl or MicrobeMethylExperiment.}
}
\value{
A numeric vector of beta values for a MicrobeMethyl object, or a
list of beta value vectors for a MicrobeMethylExperiment object.
}
\description{
This generic function retrieves the percent methylation values (beta values)
from MicrobeMethyl or MicrobeMethylExperiment objects.
}
