% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getCorrSignTable}
\alias{getCorrSignTable}
\title{Get a table in the format of 'corr (sign)'}
\usage{
getCorrSignTable(corr.sign.matrix, digits = 3)
}
\arguments{
\item{corr.sign.matrix}{The \code{\link{matrix}} of pairwised correlations and significance, 
where the upper triangle is significance and the lower triangle is correlation (or equivalent).}

\item{digits}{The number of digits to keep.}
}
\value{
The matrix of strings in the format of 'corr (sign)' for report.
}
\description{
Get a table in the format of 'corr (sign)'
}
\examples{

getCorrSignTable(corr.sign.matrix, digits=2)
}
\keyword{utils}

