% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{getDissimilarityMatrix}
\alias{getDissimilarityMatrix}
\title{dissimilarity matrix
Dissimilarity matrix of paired samples
diss.fun = "beta1-1", "jaccard", "horn.morisita"}
\usage{
getDissimilarityMatrix(expId, isPlot, min2, diss.fun = "beta1-1",
  taxa.group = "all", verbose = TRUE)
}
\description{
dissimilarity matrix
Dissimilarity matrix of paired samples
diss.fun = "beta1-1", "jaccard", "horn.morisita"
}

