% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getFirstNInString}
\alias{getFirstNInString}
\title{Return the first \code{n} elements in a string, 
using \code{\link{paste}}. 
If \code{n} is greater than length,
then return the whole vector or list in a string.}
\usage{
getFirstNInString(vect, n, collapse = ", ", tail = "...")
}
\arguments{
\item{vect}{A Vector or List.}

\item{n}{The first n elements.}
}
\description{
Return the first \code{n} elements in a string, 
using \code{\link{paste}}. 
If \code{n} is greater than length,
then return the whole vector or list in a string.
}
\examples{
printFirstN(1:10, 3)
printFirstN(1:10, 20)
}
\keyword{utils}

