% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{getMaxRemainedDiversity}
\alias{getMaxRemainedDiversity}
\title{table to max remained diversity}
\usage{
getMaxRemainedDiversity(lev.q, taxa.group = "assigned", verbose = TRUE)
}
\description{
table to max remained diversity
}

