% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOJava.R
\name{getMethodLines}
\alias{getMethodLines}
\title{Get lines containing Java method name.
The regular expression is modified from 
\url{http://stackoverflow.com/questions/68633/regex-that-will-match-a-java-method-declaration}.}
\usage{
getMethodLines(linn)
}
\arguments{
\item{linn}{A character vector of lines read from the class file. 
Refer to \code{\link{readLines}}.}
}
\value{
A character vector of lines containing Java method name.
}
\description{
Get lines containing Java method name.
The regular expression is modified from 
\url{http://stackoverflow.com/questions/68633/regex-that-will-match-a-java-method-declaration}.
}
\examples{
method.lines <- getMethodLines(linn)
}
\keyword{IO}

