% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloRareCurve.R
\name{getPhylorareDF}
\alias{getPhylorareDF}
\title{Rarefaction curve of Phylogenetic Diversity}
\usage{
getPhylorareDF(t.community.matrix, phylo.tree, min.size = 100,
  sample.sizes = NULL, points.1.min = 6, points.min.max = 9,
  log.fn = "phylo-rare.csv", verbose = T)
}
\arguments{
\item{t.community.matrix}{A transposed matrix from community matrix, 
where rows are samples, columns are OTUs.}

\item{phylo.tree}{A rooted tree of phylo object.}

\item{min.size}{return NULL, if \code{min(rowSums(t.community.matrix)) < min.size}. Default to 100.}

\item{sample.sizes}{A vector of sample sizes to build phylo rare curve. 
If NULL, then create automatically from 1 to \code{max(rowSums(t.community.matrix))}.}

\item{points.1.min, points.min.max}{If \code{sample.sizes} is NULL, the number of data points is defined 
between 1 and \code{min(rowSums(t.community.matrix))} or and \code{max(rowSums(t.community.matrix))}.}

\item{log.fn}{The CSV file name to store phylo rare data frame. If NULL, then do not log the file.}

\item{verbose}{Default to TRUE}
}
\description{
Use \code{\link{phylocurve}} in David Nipperess's package \pkg{PDcalc} 
to build rarefaction curves of phylogenetic diversity of multiple samples. 
The detail to see \url{https://github.com/davidnipperess/PDcalc} or 
\url{http://davidnipperess.blogspot.co.nz}.
}
\details{
Data input \strong{t.community.matrix} is 
a transposed matrix from community matrix we defined in \pkg{ComMA}.
\strong{phylo.tree} is a rooted tree of phylo object, 
which can get from \pkg{ape} \code{\link{read.tree}}.
}
\examples{
phylo.rare.df <- getPhylorareDF(t.community.matrix, phylo.tree)
phylo.rare.df
}

