% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{getPhyloRareTable}
\alias{getPhyloRareTable}
\title{table to plot Phylo Rarefaction}
\usage{
getPhyloRareTable(expId, isPlot, min2, taxa.group = "assigned",
  verbose = TRUE)
}
\description{
table to plot Phylo Rarefaction
}

