% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPrioritisation.R
\name{getPlotPriorByJostDiversity}
\alias{getPlotPriorByJostDiversity}
\title{Plot prioritisation by Jost diversity}
\usage{
getPlotPriorByJostDiversity(t.communityMatrix, lev, q)
}
\arguments{
\item{t.communityMatrix}{A transposed matrix from community matrix, 
where rows are plots (Use plots instead of subplots.), columns are OTUs.}

\item{lev}{Level of diversity to be calculated. Will accept: 'alpha,' 'beta,' or 'gamma.'}

\item{q}{Order of the diversity measure. Defaults to the Shannon case where q = 1.}
}
\value{
A data frame with 2 columns: \emph{rank, diversity}. Rank 1 is the most important plot, 
\emph{n} is the least important, and row.names are plot names. For example,
\tabular{rrr}{
   \tab rank \tab diversity\cr
  CM30c39 \tab 28 \tab 1845.785714\cr
  CM30c44 \tab 27 \tab 1875.888889\cr
  CM31a5 \tab 26 \tab 1899.653846 
}
}
\description{
Plot prioritisation by Jost diversity calculated from 
\pkg{vegetarian} \code{\link{d}}.
It uses a greedy algorithm to remove plots sequentially 
so as to minimize the loss of diversity among the remaining plots,
which always chooses the 1st plot if there are multi-results 
in each prioritisation loop.
Rank 1 is the most important plot and removed at the last, 
\emph{n} is the least important and removed in the beginning.
}
\examples{
t.communityMatrix <- getCommunityMatrixT("16S", isPlot=TRUE, minAbund=1, taxa.group=""BACTERIA"")
priorPlots <- getPlotPriorByJostDiversity(t.communityMatrix, lev="gamma", q=1)
}
\keyword{plot}
\keyword{prioritisation}

