% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtils.R
\name{getPlural}
\alias{getPlural}
\title{Get plural by adding 's' except special plural.}
\usage{
getPlural(...)
}
\arguments{
\item{singular}{The singular of a word in string.}
}
\value{
The plural of given singular. 
But it needs to add special plural manully in the code.
}
\description{
Get plural by adding 's' except special plural.
}
\examples{
getPlural("species")
[1] "species"
getPlural("phylumn")
[1] "phyla"
}
\keyword{utils}

