% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{getRarefactionTableTaxa}
\alias{getRarefactionTableTaxa}
\title{table to plot Rarefaction}
\usage{
getRarefactionTableTaxa(expId, isPlot, min2, taxa.group, div = "alpha1",
  verbose = TRUE)
}
\description{
table to plot Rarefaction
}

