% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphAdd.R
\name{ggAddLine}
\alias{ggAddLine}
\title{Add a line}
\usage{
ggAddLine(gg.plot, linetype = 1, xintercept, yintercept, intercept, slope,
  smooth.method, ...)
}
\arguments{
\item{gg.plot}{A \code{\link{ggplot}} object.}

\item{linetype}{Refer to \code{\link{linetype}}: 
0 = blank, 1 = solid, 2 = dashed, 3 = dotted, 
4 = dotdash, 5 = longdash, 6 = twodash.}

\item{xintercept, yintercept, intercept, slope, smooth.method}{Refer to \code{\link{geom_vline}}, \code{\link{geom_hline}}, 
\code{\link{geom_abline}}, \code{\link{geom_smooth}}. 
They cannot be used at the same time.}

\item{...}{Other arguments passed to \code{\link{layer}}.}
}
\description{
Add a line \code{\link{geom_line}} to a given \code{\link{ggplot}} object.
}
\examples{
p <- ggAddLine(p, linetype = 2, yintercept = 1)
p <- ggAddLine(p, smooth.method = "lm")
}
\keyword{graph}

