% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphAdd.R
\name{ggAddNumbers}
\alias{ggAddNumbers}
\title{Add numbers}
\usage{
ggAddNumbers(gg.plot, fun.y.lab = mean, fun.y.pos = median, y.adj = 0.98,
  digits = 2, dodge.width = 0.8, text.size = 3, text.colour = "black")
}
\arguments{
\item{gg.plot}{A \code{\link{ggplot}} object.}

\item{fun.y.lab}{A function to calculate numbers displayed in the figure.
Default to function \code{\link{mean}}. Ues \code{\link{length}} to show number of observations.}

\item{fun.y.pos}{A function to calculate the initial poistion of text on y-value. 
Default to \code{\link{median}}.}

\item{y.adj}{The propotion of the initial poistion of text on y-value. 
> 1 will raises the text, and < 1 will sinks the text. Default to 0.98.}

\item{digits}{Integer indicating the number of decimal places for \code{\link{round}}.}

\item{dodge.width}{Dodging width, when different to the width of the individual elements. 
Default to 0.8. Refer to \code{\link{position_dodge}}.}

\item{text.size}{The text size of labels. Default to 3.}

\item{text.colour}{The text colour. Default to black.}
}
\description{
Add numbers as text in a \code{\link{ggplot}} object, such as mean of box plot. 
Refer to \code{\link{stat_summary}}.
}
\examples{
p <- ggAddNumbers(p, fun.y.lab=mean)
p <- ggAddNumbers(p, fun.y.lab=length, y.adj=1.02)
}
\keyword{graph}

