% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{ggBarYAcrossX}
\alias{ggBarYAcrossX}
\title{Grouping bar chart Y across X.}
\usage{
ggBarYAcrossX(df.aggre, melt.id = "samples", print.xtable = NULL,
  title = "The number of OTUs across the number of samples",
  x.lab = "Number of samples crossed", y.lab = "Number of OTUs",
  legend.title = "", legend.labels = c("OTUs", "reads"),
  x.lab.interval = 1)
}
\arguments{
\item{df.aggre}{A data frame of row counts and sums from a community matrix. See \code{\link{rowCountAndSum}}.}

\item{print.xtable}{TRUE/FALSE to print \code{\link{xtable}} in console. 
Allow NULL, if NULL, then do not print. Default to NULL.}

\item{title}{Graph title}

\item{x.lab, y.lab}{The label of x-axis or y-axis, such as plot names.}

\item{legend.title}{The title of legend. Refer to \pkg{ggplot2} \code{\link{scale_fill_discrete}}. 
Default to a empty string.}

\item{legend.labels}{The labels of legend, which are fixed to 2 groups. Default to c("OTUs", "reads").}

\item{x.lab.interval}{The x labels interval. Default to 1 to dispay lables for every values.}
}
\description{
Create a grouping bar chart given community matrix to display 
the number of OTUs (y-axis) across the number of samples (x-axis). 
The 'red' bar is the number of OTUs appeared only in that number of samlpes, 
and the 'green' bar is the number of reads assigned to those OTUs.
}
\examples{
communityMatrix <- getCommunityMatrix("16S", isPlot=TRUE, minAbund=1)
cm.aggre <- cmYAcrossX(communityMatrix)
gg.plot <- ggBarYAcrossX(cm.aggre)
pdfGgplot(gg.plot, fig.path="bar-otus-across-samples.pdf", width=8, height=8)  
}
\keyword{graph}

