% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{ggNMDSPlot}
\alias{ggNMDSPlot}
\title{Use \code{\link{metaMDS}} in \code{\link{vegan}}
to create Nonmetric Multidimensional Scaling (NMDS) plot .}
\usage{
ggNMDSPlot(comm, distance = "bray", ..., title = "MDS", x.lab = "MDS1",
  y.lab = "MDS2")
}
\arguments{
\item{comm}{Community data for NMDS plot. It is dissimilarities either 
as a \code{\link{dist}} structure or as a symmetric square \code{\link{matrix}}.
Refer to \code{\link{metaMDS}}.}

\item{distance}{Dissimilarity index used in \code{\link{vegdist}}.}

\item{attr.df}{A data frame to define how NMDS plot is coloured, shaped, or linked.
where its row names have to match row names of community data \code{comm}.}
}
\description{
Use \code{\link{metaMDS}} in \code{\link{vegan}}
to create Nonmetric Multidimensional Scaling (NMDS) plot .
}
\examples{

nmds.plot <- ggNMDSPlot(comm, attr.df, coloured.by="", shaped.by="", linked.by="")
}
\keyword{graph}

