% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{ggPercentageBarChart}
\alias{ggPercentageBarChart}
\title{Percentage bar chart coloured by groups.}
\usage{
ggPercentageBarChart(df, melt.id, title = "Percent Bar Chart", x.lab = "",
  y.lab = "", autoWidth = TRUE)
}
\arguments{
\item{df}{A data frame to \code{\link{melt}} and then make a percent bar chart. 
For example,
\tabular{rrrrr}{
  Phyla \tab 16s \tab 18s \tab ITS \tab TaxaGroup\cr
  Actinobacteria \tab 958 \tab 1 \tab 3 \tab Bacteria\cr
  Crenarchaeota \tab 1 \tab 0 \tab 0 \tab Archaea\cr
  Ascomycota \tab 2 \tab 765 \tab 971 \tab Fungi 
}}

\item{melt.id}{A column name to \code{\link{melt}} and used to assign the colours.}

\item{title}{Graph title}

\item{x.lab, y.lab}{The label of x-axis or y-axis, such as plot names.}

\item{autoWidth}{If TRUE, then use number of bars and legend columns 
to estimate pdf width automatically. Default to TRUE.}

\item{low, }{high Refer to \pkg{ggplot2} \code{\link{scale_fill_gradient}}. 
Default to low="white", high="steelblue".}
}
\description{
Percentage bar chart coloured by groups.
}
\examples{
taxa.phyla <- readFile("./data/examples/taxonomy97phyla.txt")
bar.chart <- ggPercentageBarChart(taxa.phyla, melt.id="TaxaGroup")
pdfGgplot(bar.chart$gg.plot, fig.path="taxa-percentage-bar.pdf", width=bar.chart$pdf.width, height=8) 
}
\keyword{graph}

