% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{ggScatterPlotEllipse}
\alias{ggScatterPlotEllipse}
\title{Data ellipse on a scatter plot.}
\source{
\url{http://stackoverflow.com/questions/2397097/how-can-a-data-ellipse-be-superimposed-on-a-ggplot2-scatterplot}.
}
\usage{
ggScatterPlotEllipse(df.clusters, title = "Clusters", point.size = 3,
  palette = "Set1", addLabel = FALSE, label.size = 3, hjust = -0.1,
  vjust = -0.2, alpha = 0.5)
}
\arguments{
\item{df.clusters}{The 3-column data frame for plotting. 
The first 2 columns are coordinates, and 3rd is cluster names.}

\item{point.size}{The size of points from \code{\link{point.size}}. Default to 3.}

\item{palette}{The palette to colour clusters using \code{\link{scale_colour_brewer}}. 
Default to 'Set1' (max 8 colours). Refer to \url{http://www.datavis.ca/sasmac/brewerpal.html}.}

\item{addLabel}{Add row names to scatter plot. Default to FALSE}
}
\value{
A \code{\link{gtable}} object of scatter plot.
}
\description{
Superimposed data ellipse on a \pkg{ggplot2} scatter plot.
}
\examples{
df.clusters <- random2Clusters()
df.clusters
g.table <- ggScatterPlotEllipse(df.clusters, addLabel=T)
pdfGtable(g.table, fig.path="clusters-scatter-plot.pdf") 
}
\keyword{graph}

