% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Share a legend between multiple plots using \code{\link{grid.arrange}}.}
\source{
\url{http://rpubs.com/sjackman/grid_arrange_shared_legend}
}
\usage{
grid_arrange_shared_legend(...)
}
\arguments{
\item{...}{The list of \code{\link{ggplot}} objects.}
}
\value{
A \code{\link{gtable}} object of \code{\link{grid.arrange}}.
}
\description{
Share a legend between multiple plots using \code{\link{grid.arrange}}.
}
\examples{
library(ggplot2); library(grid); library(gridExtra)
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
p1 <- qplot(carat, price, data=dsamp, colour=clarity)
p2 <- qplot(cut, price, data=dsamp, colour=clarity)
p3 <- qplot(color, price, data=dsamp, colour=clarity)
p4 <- qplot(depth, price, data=dsamp, colour=clarity)
grid_arrange_shared_legend(p1, p2, p3, p4)
}
\keyword{utils}

