% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{gtPCAPlot}
\alias{gtPCAPlot}
\title{PCA Plot}
\usage{
gtPCAPlot(comm, attr.df, x.i = 1, y.i = 2, colour.id = NULL,
  shape.id = NULL, linke.id = NULL, add.text = TRUE, text.data = NULL,
  text.size = 3, title = "PCA", verbose = TRUE, ...)
}
\arguments{
\item{comm}{Community data for NMDS plot. 
It is dissimilarities either as a \code{\link{dist}} structure 
or as a community data defined by \code{\link{vegan}} which is 
a transposed matrix of community matrix in \code{\link{ComMA}}.
Use \code{\link{transpose.df}} to rotate the data frame.
Detail to \code{\link{prcomp}}.}

\item{attr.df}{A data frame to define how NMDS plot is coloured, shaped, or linked.
where its row names have to match row names of community data \code{comm}.}
}
\description{
Use \code{\link{prcomp}} to create Principal Components Analysis (PCA) plot.
It is extended from \code{\link{gtScatterPlot}}.
}
\examples{

pca.plot <- gtNMDSPlot(comm, env, colour.id="FishSpecies", shape.id="FeedingPattern", add.text=T)
require(grid)
grid.draw(pca.plot)
}
\keyword{graph}

