% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{gtScatterPlotEllipse}
\alias{gtScatterPlotEllipse}
\title{Superimposed data ellipse on a scatter plot, which is extended from \code{\link{ggScatterPlot}}.}
\usage{
gtScatterPlotEllipse(df.clusters, title = "Clusters", palette = "Set1", ...)
}
\arguments{
\item{df.clusters}{The 3-column data frame for plotting. 
The first 2 columns are coordinates, and 3rd is cluster names.}

\item{palette}{The palette to colour clusters using \code{\link{scale_colour_brewer}}. 
Default to 'Set1' (max 8 colours). Refer to \url{http://www.datavis.ca/sasmac/brewerpal.html}.}
}
\value{
A \code{\link{gtable}} object of scatter plot.
}
\description{
Superimposed data ellipse on a scatter plot, which is extended from \code{\link{ggScatterPlot}}.
}
\examples{
df.clusters <- random2Clusters()
df.clusters
g.table <- gtScatterPlotEllipse(df.clusters, add.text=T)
pdfGtable(g.table, fig.path="clusters-scatter-plot.pdf") 
}
\keyword{graph}

