% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOLog.R
\name{IOLog}
\alias{IOLog}
\alias{readBEASTLog}
\title{IO functions to read BEAST log}
\usage{
readBEASTLog(file, verbose = TRUE)
}
\arguments{
\item{file}{The file to read/write.}
}
\description{
IO functions to read BEAST 1 \url{http://beast.bio.ed.ac.uk} 
or BEAST 2 \url{http://www.beast2.org} log files.
}
\details{
\code{readBEASTLog} reads BEAST log files and return a data frame 
whose column names are parameters and row names are samples.
}
\examples{
beast.log <- readBEASTLog("data-raw/star.beast.log")

}
\keyword{IO}

