% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{merge2DF}
\alias{merge2DF}
\title{Merge two data frames or matrices with a same structure in one, 
put all values in 2nd data frame into brackets.}
\usage{
merge2DF(df, df2, merge.zero = TRUE, return.df = TRUE,
  stringsAsFactors = FALSE, check.names = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{df2}{The 2nd data frame whose values are into brackets.}

\item{merge.zero}{Default to TRUE to remove all " (0)".}

\item{return.df}{Default to TRUE to return a data frame, otherwise a matrix.}
}
\description{
Merge two data frames or matrices with a same structure in one, 
put all values in 2nd data frame into brackets.
}
\examples{
df <- merge2DF(df, df2)
}
\keyword{utils}

