% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{mergeByRownames}
\alias{mergeByRownames}
\title{Merge two data frames by 'row.names' using \code{\link{merge}}.}
\usage{
mergeByRownames(x, y, warning.msg = TRUE, ...)
}
\arguments{
\item{x, y}{data frames, or objects to be coerced to one.}

\item{warning.msg}{logical; if TRUE, then print warning message when rows are missing after merge.}

\item{...}{pass to \code{\link{merge}}.}
}
\description{
Merge two data frames by 'row.names' using \code{\link{merge}}.
}
\examples{
df <- combineTwoDF(df, df2)
}
\keyword{utils}

