% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{methylation}
\alias{methylation}
\alias{methylation,commaData-method}
\title{Accessor for the methylation (beta value) matrix}
\usage{
methylation(object)

\S4method{methylation}{commaData}(object)
}
\arguments{
\item{object}{A \code{commaData} object.}
}
\value{
A numeric matrix with rows corresponding to methylation sites and
columns corresponding to samples. Rownames are site keys
(\code{"chrom:position:strand:mod_type"}); column names are sample names.
}
\description{
Retrieves the sites × samples matrix of methylation beta values from a
\code{\link{commaData}} object. Values are in the range 0–1 (proportion
of reads called methylated). Sites below the \code{min_coverage} threshold
set at object creation are \code{NA}.
}
\examples{
data(comma_example_data)
m <- methylation(comma_example_data)
dim(m)
head(m)

}
\seealso{
\code{\link[=coverage,commaData-method]{coverage}}, \code{\link{siteInfo}},
\code{\link{sampleInfo}}
}
