% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{methylBtoM}
\alias{methylBtoM}
\title{Convert Methylation B-values to M-values}
\usage{
methylBtoM(B, alpha = 0.001, sym = FALSE, base = 2)
}
\arguments{
\item{B}{a numeric vector of Beta-values.}

\item{alpha}{a double. Used to prevent division by zero and should be an
order of magnitude smaller than your smallest beta value.}

\item{sym}{a boolean. If \code{TRUE}, use the symmetrical logarithm \code{\link[=symlog]{symlog()}}. If
\code{FALSE}, use base function \code{\link[=log]{log()}}.}

\item{base}{a positive numeric. The base with respect to which logarithms
are computed. Ultimately passed to \code{\link[=log]{log()}}. Don't change this.}
}
\value{
a numeric vector of M-values
}
\description{
Convert Methylation B-values to M-values
}
\examples{
# Sample data to make a mock methylation sample.
data <- sample(all_samples_long$beta, size = 10000, replace = TRUE)

# Use default log2 when all beta values fall between 0 and 1.
beta <- methylBtoM(data)

# Use symmetrical log2 when values fall outside 0 and 1.
beta <- methylBtoM(data, sym = TRUE)
}
