% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{methylMtoB}
\alias{methylMtoB}
\title{Title Convert Methylation M-values to B-values}
\usage{
methylMtoB(M)
}
\arguments{
\item{M}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Title Convert Methylation M-values to B-values
}
\examples{
data <- sample(all_samples_long$beta, size = 10000, replace = TRUE)
mvalues <- methylBtoM(data)
bvalues <- methylMtoB(mvalues)
}
