% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylRollingMean.R
\name{methylRollingMean}
\alias{methylRollingMean}
\title{Calculate rolling mean of methylation levels}
\usage{
methylRollingMean(
  df,
  position_col,
  methyl_col,
  w_size = 1000,
  genome_size = 4641652,
  verbose = FALSE
)
}
\arguments{
\item{df}{A data frame with columns for genomic position and methylation level.}

\item{position_col}{The name of the column in \code{df} that contains the genomic position data.}

\item{methyl_col}{The name of the column in \code{df} that contains the methylation level data.}

\item{w_size}{The window size for calculating the rolling mean. Default value is 1000.}

\item{genome_size}{The size of the genome. Default value is 4641652.}

\item{verbose}{If \code{TRUE}, progress updates will be printed. Default value is \code{FALSE}.}
}
\value{
A data frame with columns for genomic position and the rolling mean methylation level for that position.
}
\description{
Calculates the rolling mean of methylation levels for a given window size. The function takes in a data frame with columns
for genomic position and methylation level, and returns a data frame with columns for position and the rolling mean
methylation level for that position.
}
\examples{
df <- data.frame(position = c(1, 2, 3, 4, 5), methyl = c(0.1, 0.2, 0.3, 0.4, 0.5))
result <- methylRollingMean(df, position_col = "position", methyl_col = "methyl", w_size = 2)

}
