% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl_classes.R
\docType{class}
\name{MicrobeMethyl-class}
\alias{MicrobeMethyl-class}
\alias{MicrobeMethyl}
\title{MicrobeMethyl class}
\description{
An S4 class to represent adenine methylation (6mA) data from a single
bacterial sample, particularly in \emph{E. coli}. This class stores methylation
data (beta values and number of reads), sample name, and sample metadata.
}
\section{Slots}{

\describe{
\item{\code{sample_name}}{A character string representing the name of the sample.}

\item{\code{assembly}}{A character vector with the chromosome name(s) contained in
the object.}

\item{\code{sample_metadata}}{A list containing metadata for the sample, such as
experimental conditions and any other relevant information.}

\item{\code{site_metadata}}{A data.frame containing additional site-level annotation
that accompanies the methylation measurements.}
}}

