% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{MicrobeMethyl}
\alias{MicrobeMethyl}
\title{MicrobeMethyl constructor}
\usage{
MicrobeMethyl(assembly, start, end, coverage, percentMethylation, sample_name)
}
\arguments{
\item{assembly}{A character string indicating the chromosome name.}

\item{start}{A numeric vector with start positions for each methylated base.}

\item{end}{A numeric vector with end positions for each methylated base.}

\item{coverage}{A numeric vector with coverage values for each position.}

\item{percentMethylation}{A double vector with percent methylation
(in decimals 0-1) for each position.}

\item{sample_name}{A character string indicating the sample name.}
}
\value{
A MicrobeMethyl object
}
\description{
This is a constructor function for the
\code{\linkS4class{MicrobeMethyl}} class.
}
\examples{
microbe_methyl_obj <- MicrobeMethyl(assembly = "NC_000913.3",
                                    start = c(620, 621, 727, 728),
                                    end = c(620, 621, 727, 728),
                                    coverage = c(41, 89, 110, 76),
                                    percentMethylation = c(0.99, 0.95, 1, 0.84),
                                    sample_name = "WT1")

}
