% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\docType{class}
\name{MicrobeMethylExperiment-class}
\alias{MicrobeMethylExperiment-class}
\alias{MicrobeMethylExperiment}
\title{MicrobeMethylExperiment class}
\description{
An S4 class to represent an experiment containing multiple MicrobeMethyl
objects. This class stores a list of MicrobeMethyl objects and experiment
metadata.
}
\section{Slots}{

\describe{
\item{\code{samples}}{A list of MicrobeMethyl objects, each representing a single
sample from the same experiment.}

\item{\code{experiment_metadata}}{A list containing metadata for the experiment,
such as experimental conditions, date, and any other relevant information.}
}}

