% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{MicrobeMethylExperiment}
\alias{MicrobeMethylExperiment}
\title{MicrobeMethylExperiment constructor}
\usage{
MicrobeMethylExperiment(..., experiment_metadata)
}
\arguments{
\item{...}{One or more MicrobeMethyl objects.}

\item{experiment_metadata}{A list containing metadata for the experiment,
such as experimental conditions, date, and any other relevant information.}
}
\value{
A MicrobeMethylExperiment object
}
\description{
This is a constructor function for the
\code{\linkS4class{MicrobeMethylExperiment}} class.
}
\examples{
microbe_methyl_obj1 <- MicrobeMethyl(assembly = "NC_000913.3",
                                     start = c(620, 621, 727, 728),
                                     end = c(620, 621, 727, 728),
                                     coverage = c(41, 89, 110, 76),
                                     percentMethylation = c(0.99, 0.95, 1, 0.84),
                                     sample_name = "WT1")
microbe_methyl_obj2 <- MicrobeMethyl(assembly = "NC_000913.3",
                                     start = c(630, 631, 737, 738),
                                     end = c(630, 631, 737, 738),
                                     coverage = c(45, 90, 100, 80),
                                     percentMethylation = c(0.98, 0.92, 1, 0.82),
                                     sample_name = "WT2")
experiment_metadata <- list(date = "2023-05-02", condition = "wild_type")
experiment_obj <- MicrobeMethylExperiment(microbe_methyl_obj1, microbe_methyl_obj2,
                                          experiment_metadata = experiment_metadata)

}
