% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{mkdir}
\alias{mkdir}
\title{Create folder from \code{\link{file.path}} if not exist.}
\usage{
mkdir(dir.path)
}
\arguments{
\item{dir.path}{The folder path constructed by \code{\link{file.path}}.}
}
\description{
Create folder from \code{\link{file.path}} if not exist.
}
\examples{
figDir <- file.path(workingPath, "figures")
mkdir(figDir) 
}
\keyword{IO}

