% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{mvRowsToLast}
\alias{mvRowsToLast}
\title{Move rows to the last rows of data frame given a column value}
\usage{
mvRowsToLast(data, col, regex = "unclassified", ignore.case = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{col}{The column, which can be a string (column name) 
or number (index of the column).}

\item{regex}{Regular expression for \code{\link{grep}} to indetify rows. 
Default to "unclassified". Use "^string$" to perform exact match.}

\item{ignore.case}{Refer to \code{\link{grep}}.}
}
\value{
The data frame of rows moved to the last.
}
\description{
Move rows to the last rows of data frame given a column value
}
\examples{

# move all rows having "unclassified" phyla to the last rows
data <- mvRowsToLast(data, "phylum")
# move all rows having exctly mathced "Bacteria" in column 1 to the last rows
data <- mvRowsToLast(data, 1, regex="^Bacteria$")
}
\keyword{utils}

