% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCHS2022_sample.R
\docType{data}
\name{NCHS2022_sample}
\alias{NCHS2022_sample}
\title{Example data from the National Vital Statistics System of the National Center for Health Statistics (NCHS), 2022}
\format{
A dataframe 30 columns, including demographic and birth information
for a random sample of 20,000 singleton births from nulliparous mothers in the
US in 2022.
}
\source{
\url{https://data.nber.org/nvss/natality/inputs/raw/2022/}
}
\usage{
NCHS2022_sample
}
\description{
Example data from the National Vital Statistics System of the National Center for Health Statistics (NCHS), 2022
}
\examples{
\dontrun{
data("NCHS2022_sample")
head(NCHS2022_sample)
}
}
\keyword{datasets}
