% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{normalize}
\alias{normalize}
\title{normalize given vector.}
\source{
\url{https://stat.ethz.ch/pipermail/r-help//2012-October/336676.html}.
}
\usage{
normalize(vec)
}
\arguments{
\item{vec}{A vector.}
}
\description{
normalize given vector.
}
\examples{
normed <- as.data.frame(lapply(mtcars, normalize))
lapply(normed, range)
}
\keyword{utils}

