% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeMethylation.R
\name{normalizeMethylation}
\alias{normalizeMethylation}
\title{Title Normalize Methylation Values}
\usage{
normalizeMethylation(
  df,
  alpha = 0.001,
  normalize_position = TRUE,
  rescale = FALSE,
  plots = TRUE
)
}
\arguments{
\item{df}{a data frame containing methylation, position, and coverage values.}

\item{plots}{}
}
\description{
Title Normalize Methylation Values
}
