% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBFFio.R, R/parseGenBankFeatures.R
\name{parseGenBankFeatures}
\alias{parseGenBankFeatures}
\title{Parse GenBank Features into GRanges}
\usage{
parseGenBankFeatures(features_lines, seqname)

parseGenBankFeatures(features_lines, seqname)
}
\arguments{
\item{features_lines}{A character vector containing the lines of the FEATURES section.}

\item{seqname}{The sequence name (e.g., accession number).}
}
\value{
A \code{GRanges} object with genomic features.

A \code{GRanges} object with genomic features.
}
\description{
This function parses the FEATURES section of a GenBank file and converts it into a \code{GRanges} object.

This function parses the FEATURES section of a GenBank file and converts it into a \code{GRanges} object.
}
