% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBFFio.R
\name{parseGenBankSequence}
\alias{parseGenBankSequence}
\title{Parse GenBank Sequence}
\usage{
parseGenBankSequence(sequence_lines)
}
\arguments{
\item{sequence_lines}{A character vector containing the lines of the ORIGIN section.}
}
\value{
A character string representing the sequence.
}
\description{
This function parses the sequence from the ORIGIN section of a GenBank file.
}
