% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBFFio.R
\name{parseLocationString}
\alias{parseLocationString}
\title{Parse Location String from GenBank Feature}
\usage{
parseLocationString(location_string)
}
\arguments{
\item{location_string}{A character string representing the location.}
}
\value{
A list containing:
\item{ranges}{An \code{IRanges} object with the feature ranges.}
\item{strand}{A character representing the strand ("+", "-", or "*").}
}
\description{
This function parses a location string from a GenBank feature and returns ranges and strand information.
}
