% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{pdf}
\alias{grid_arrange_shared_legend}
\alias{pdf}
\alias{pdf.ggplot}
\alias{pdf.gtable}
\alias{pdf.plot}
\title{Create pdf}
\source{
\code{grid_arrange_shared_legend} is derived from 
\url{http://rpubs.com/sjackman/grid_arrange_shared_legend}
}
\usage{
pdf.ggplot(gg.plot, fig.path, width = 6, height = 6, useDingbats = TRUE)

pdf.gtable(g.table, fig.path, width = 6, height = 6, useDingbats = TRUE)

pdf.plot(plot, fig.path, width = 6, height = 6, useDingbats = TRUE)

grid_arrange_shared_legend(...)
}
\arguments{
\item{gg.plot}{A \code{\link{ggplot}} object.}

\item{fig.path}{fig.path The full path of pdf file.}

\item{width, }{height Refer to \code{\link{pdf}}. Default to width=6, height=6.}

\item{useDingbats}{Defaults to TRUE, which produces smaller and better output. 
Setting this to FALSE can work around font display problems in broken PDF viewers.
Refer to \code{\link{pdf}}.}

\item{plot}{The list returned from \code{\link{plot}}.}

\item{...}{The list of \code{\link{ggplot}} objects.}

\item{gtable}{A \code{\link{gtable}} object.}
}
\details{
\code{pdf.ggplot} creates pdf for ggplot object.

\code{pdf.gtable} creates pdf for gtable object.

\code{pdf.plot} creates pdf for the list returned from \code{\link{plot}}.

\code{grid_arrange_shared_legend} shares a legend 
between multiple plots using \code{\link{grid.arrange}}. 
Return a \code{\link{gtable}} object.
}
\examples{
pdf.ggplot(gg.plot, fig.path="fig.pdf", width=8, height=8)  

pdf.gtable(g.table, fig.path="fig.pdf", width=8, height=8)  

pdf.plot(plot, fig.path="fig.pdf", width=8, height=8)  

library(ggplot2); library(grid); library(gridExtra)
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
p1 <- qplot(carat, price, data=dsamp, colour=clarity)
p2 <- qplot(cut, price, data=dsamp, colour=clarity)
p3 <- qplot(color, price, data=dsamp, colour=clarity)
p4 <- qplot(depth, price, data=dsamp, colour=clarity)
grid_arrange_shared_legend(p1, p2, p3, p4)

}
\keyword{graph}

