% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{pdfGgplot}
\alias{pdfGgplot}
\title{Create pdf for ggplot object.}
\usage{
pdfGgplot(gg.plot, fig.path, width = 6, height = 6)
}
\arguments{
\item{gg.plot}{A \code{\link{ggplot}} object.}

\item{fig.path}{fig.path The full path of pdf file.}

\item{width, }{height Refer to \code{\link{pdf}}. Default to width=6, height=6.}
}
\description{
Create pdf for ggplot object.
}
\examples{
pdfGgplot(gg.plot, fig.path="fig.pdf", width=8, height=8)  
}
\keyword{utils}

