% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphUtils.R
\name{pdfGtable}
\alias{pdfGtable}
\title{Create pdf for gtable object.}
\usage{
pdfGtable(g.table, fig.path, width = 8, height = 8)
}
\arguments{
\item{fig.path}{fig.path The full path of pdf file.}

\item{width, }{height Refer to \code{\link{pdf}}. Default to width=6, height=6.}

\item{gtable}{A \code{\link{gtable}} object.}
}
\description{
Create pdf for gtable object.
}
\examples{
pdfGtable(g.table, fig.path="fig.pdf", width=8, height=8)  
}
\keyword{utils}

