% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPrioritisation.R
\name{PlotPrioritisation}
\alias{PlotPrioritisation}
\alias{getPlotPrior.JostDiver}
\title{Plot prioritisation}
\usage{
getPlotPrior.JostDiver(t.community.matrix, lev = c("alpha", "beta", "gamma"),
  q)
}
\arguments{
\item{t.community.matrix}{A transposed matrix from community matrix, 
where rows are plots (Use plots instead of subplots.), columns are OTUs.}

\item{lev}{Level of diversity to be calculated. Will accept: 'alpha,' 'beta,' or 'gamma.'}

\item{q}{Order of the diversity measure. Defaults to the Shannon case where q = 1.}
}
\value{
A data frame with 2 columns: \emph{rank, diversity}. Rank 1 is the most important plot, 
\emph{n} is the least important, and row.names are plot names. For example,
\tabular{rrr}{
   \tab rank \tab diversity\cr
  CM30c39 \tab 28 \tab 1845.785714\cr
  CM30c44 \tab 27 \tab 1875.888889\cr
  CM31a5 \tab 26 \tab 1899.653846 
}
}
\description{
Ranking of sample plots by their contributions 
to the total biodiversity.
}
\details{
Plot prioritisation by Jost diversity calculated from 
\pkg{vegetarian} \code{\link{d}}.
It uses a greedy algorithm to remove plots sequentially 
so as to minimize the loss of diversity among the remaining plots,
which always chooses the 1st plot if there are multi-results 
in each prioritisation loop.
Rank 1 is the most important plot and removed at the last, 
\emph{n} is the least important and removed in the beginning.
}
\examples{
plot.prior.g1 <- getPlotPrior.JostDiver(t.community.matrix, lev="gamma", q=1)

}
\keyword{plot}
\keyword{prioritisation}

