% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{prettyNumbers}
\alias{prettyNumbers}
\title{\code{prettyNumbers} provide pretty numbers with comma separator 
to a given data frame \code{df}.}
\usage{
prettyNumbers(df, digits = 2, drop.0.tail = TRUE)
}
\arguments{
\item{digits, }{integer indicating the number of decimal places, default o 2.}

\item{drop.0.tail}{If TRUE, then remove all 0's decimal, such as ".00"}
}
\description{
\code{prettyNumbers} provide pretty numbers with comma separator 
to a given data frame \code{df}.
}
\examples{
prettyNumbers(df)
}

