% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{printXTable}
\alias{printXTable}
\title{Print data fram (table) as Latex format to either file or console.}
\usage{
printXTable(df, caption, label, include.rownames = TRUE, file = NULL,
  align = NULL, digits = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{caption}{Latex table caption.}

\item{label}{Latex table label.}

\item{file}{If NULL, then print the results to console, otherwise print them to the file. Default to NULL.}

\item{align}{Refer to \code{\link{xtable}}.}

\item{digits}{Refer to \code{\link{xtable}}.}
}
\description{
Print data fram (table) as Latex format to either file or console.
}
\examples{
tableFile <- file.path(workingPath, "report.tex")
printXTable(data.frame, caption = "Phylogenetic beta diversity", 
            label = "tab:pd:beta", file=tableFile)
}
\keyword{IO}

