% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedundancyAnalysis.R
\name{proceedRDA}
\alias{proceedRDA}
\title{Redundancy Analysis using \pkg{vegan} \code{\link{capscale}}}
\usage{
proceedRDA(cm.or.dist, env, matrix.name = "", taxa.group = "",
  table.file = NULL, verbose = TRUE)
}
\arguments{
\item{cm.or.dist}{A data frame or dist. Rows are samples.}

\item{env}{The enviornmental data, where rows are samples, and must be same as rownames(cm.or.dist) inlcuding order.}

\item{matrix.name}{The string to locate the matrix from its file name. 
Only used for table name and label here.}

\item{taxa.group}{The taxonomic group. Only used for table name and label here.}

\item{table.file}{If NULL, then print the results to console, 
otherwise print them to the file. Default to NULL.}

\item{verbose}{More details. Default to TRUE.}
}
\value{
A list of results from RDA including 3 data frames.
}
\description{
Constrained Analysis of Principal Coordinates for eDNA data set 
given environmental variables.
}
\examples{
rda.list <- proceedRDA(cm.or.dist, env, matrix.name="16S", taxa.group="BACTERIA")
}

