% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{random2Clusters}
\alias{random2Clusters}
\title{Generate coordinates for 2 clusters.}
\source{
\url{http://stackoverflow.com/questions/2397097/how-can-a-data-ellipse-be-superimposed-on-a-ggplot2-scatterplot}.
}
\usage{
random2Clusters(n = 100, seed = 101)
}
\arguments{
\item{n}{The number of points. Default to 100.}

\item{seed}{An integer seed for \code{\link{set.seed}}. Default to 101.}
}
\description{
Generate coordinates for 2 clusters.
}
\examples{
df.clusters <- random2Clusters()
}
\keyword{utils}

