% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedundancyAnalysis.R
\name{RDA}
\alias{RDA}
\alias{printXTable.RDA}
\alias{proceedRDA}
\title{Redundancy Analysis}
\usage{
proceedRDA(cm.or.dist, env, verbose = TRUE)

printXTable.RDA(rda, table.file = NULL, invalid.char = FALSE,
  matrix.name = "", taxa.group = "")
}
\arguments{
\item{cm.or.dist}{A data frame or dist. Rows are samples.}

\item{env}{The enviornmental data, where rows are samples, 
and must be same as rownames(cm.or.dist) inlcuding order.}

\item{verbose}{More details. Default to TRUE.}

\item{rda}{The list of results from \code{proceedRDA}.}

\item{table.file}{If NULL, then print the results to console, 
otherwise print them to the file. Default to NULL.}

\item{matrix.name}{The string to locate the matrix from its file name. 
Only used for table name and label here.}

\item{taxa.group}{The taxonomic group. Only used for table name and label here.}
}
\description{
Redundancy Analysis using \pkg{vegan} 
\code{\link{capscale}}.
}
\details{
\code{proceedRDA} makes Constrained Analysis of Principal Coordinates for 
eDNA data sets given environmental variables.

\code{printXTable.RDA} prints \code{\link{xtable}} given rda results.
}
\examples{
rda <- proceedRDA(cm.or.dist, env)

printXTable.RDA(rda, table.file=NULL, matrix.name="16S", taxa.group="BACTERIA")

}
\keyword{rda}

