% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHtml.R
\name{readCodeCoverage}
\alias{readCodeCoverage}
\title{Use \code{\link{readHTMLTable}} to covert the report of code coverage 
into a list of two data frames or matrices of percentages.}
\usage{
readCodeCoverage(doc, ...)
}
\arguments{
\item{doc}{HTML document which can be a file name or a URL or 
an already parsed HTMLInternalDocument, 
or an HTML node of class XMLInternalElementNode, 
or a character vector containing the HTML content to parse and process.}

\item{...}{Other arguments passed to \code{\link{readHTMLTable}}.}
}
\value{
A list of two data frames or matrices, the 1st is Package, 2nd is Class.
}
\description{
Use \code{\link{readHTMLTable}} to covert the report of code coverage 
into a list of two data frames or matrices of percentages.
}
\examples{
code.coverage <- readCodeCoverage("code_coverage/index.html")
# $Package
#     Package Class \% Method \% Line \%
# 1    all classes    38.2     26.1   26.6
# $Class
#     Package Class \% Method \% Line \%
# 1    beast.app    36.4      8.9      1
# 2    beast.app.beastapp      20      2.7    0.5
}
\keyword{IO}

