% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readFile}
\alias{readFile}
\title{Read a file to a data frame.}
\usage{
readFile(file, sep = "\\t", header = TRUE, row.names = 1,
  verbose = TRUE, msg.file = "file", msg.col = "columns",
  msg.row = "rows", ...)
}
\arguments{
\item{file}{The 1st row is column names, the 1st column is row names.}

\item{sep}{Only used for non \emph{csv} file. Default to tab "\\t".}

\item{header}{A logical value indicating whether the file contains 
the names of the variables as its first line. Default to TRUE.}

\item{row.names}{A vector of row names. This can be a vector giving the actual row names, 
or a single number giving the column of the table which contains the row names, 
or character string giving the name of the table column containing the row names.
If there is a header and the first row contains one fewer field than the number of columns, 
the first column in the input is used for the row names. 
Otherwise if row.names is NULL, the rows are numbered. Default to 1.
Using row.names = NULL forces row numbering.}

\item{verbose}{More details. Default to TRUE.}

\item{msg.file, msg.col, msg.row}{The message regarding file, column and row, if verbose=TRUE.}

\item{...}{Other arguments passed to \code{\link{read.table}}.}
}
\value{
A data frame from the file, such as
\tabular{rrrr}{
  OTU_id \tab plot01 \tab plot02\tab ...\cr
  OTU_1 \tab 1 \tab 0 \tab ...\cr
  OTU_2 \tab 100 \tab 200 \tab ...\cr
  OTU_3 \tab 56 \tab 3 \tab ...
}
}
\description{
If the file extension is \emph{csv}, 
then use \code{\link{read.csv}}, otherwise use \code{\link{read.table}}.
}
\examples{
community.matrix <- readFile("16S.txt", msg.file="16S OTU table", msg.col="samples", msg.row="OTUs")
taxa.table <- readFile("16S_taxonomy_table.txt", msg.file="16S taxonomy table", msg.row="OTUs")
env <- readFile("env_data.txt", msg.file="enviornmental data", msg.row="samples")
taxa.phyla <- readFile("taxonomy97phyla.txt", row.names=NULL)
}
\keyword{IO}

