% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOLog.R
\name{readMCMCLog}
\alias{readMCMCLog}
\title{Reads MCMC log files and return a data frame
whose column names are parameters and row names are the number
of states at each sample.}
\usage{
readMCMCLog(file, rm.na.col = TRUE, ...)
}
\arguments{
\item{file}{The file to read/write.}

\item{rm.na.col}{If TRUE, then remove all columns with all
missing values (NA). Default to TRUE.}

\item{...}{Other arguments passed to \code{\link{readFile}}.}
}
\description{
Reads MCMC log files and return a data frame
whose column names are parameters and row names are the number
of states at each sample.
}
\examples{
mcmc.log <- readMCMCLog("data/star.beast.log")
}
\keyword{IO}

