% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversities.R
\name{richnessPerSample}
\alias{richnessPerSample}
\title{richness (OTUs/species) per sample
return 1-column data frame}
\usage{
richnessPerSample(t.community.matrix, hasTotal = TRUE)
}
\description{
richness (OTUs/species) per sample
return 1-column data frame
}

