% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{sampleNames,MicrobeMethylExperiment-method}
\alias{sampleNames,MicrobeMethylExperiment-method}
\title{Extract sample names from a MicrobeMethylExperiment object}
\usage{
\S4method{sampleNames}{MicrobeMethylExperiment}(object)
}
\arguments{
\item{object}{A MicrobeMethylExperiment object.}
}
\value{
A character vector of sample names.
}
\description{
This method extracts the sample names from a
\code{\linkS4class{MicrobeMethylExperiment}} object.
}
\examples{
# Using the example MicrobeMethylExperiment object created earlier
sample_names <- sampleNames(experiment_obj)
print(sample_names)

}
