% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{sampleNames}
\alias{sampleNames}
\title{Generic function for sampleNames}
\usage{
sampleNames(object)
}
\arguments{
\item{object}{An object for which to extract sample names.}
}
\value{
A character vector of sample names.
}
\description{
Generic function for sampleNames
}
