% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversities.R
\name{shannonPerSample}
\alias{shannonPerSample}
\title{Shannon index (gamma1) per sample
return 1-column data frame}
\usage{
shannonPerSample(t.community.matrix, digits = 2)
}
\description{
Shannon index (gamma1) per sample
return 1-column data frame
}

