% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{splitPath}
\alias{splitPath}
\title{Split a file path into folder names vector}
\source{
\url{http://stackoverflow.com/questions/29214932/split-a-file-path-into-folder-names-vector}
}
\usage{
splitPath(path)
}
\arguments{
\item{path}{The file path to be split.}
}
\value{
A the reversed vector of folder names.
}
\description{
Split a file path into folder names vector
}
\examples{
split_path("/home/foo/stats/index.html")
[1] "index.html" "stats"      "foo"        "home"      
split_path("C:\\\\Windows\\\\System32")
[1] "System32" "Windows"  "C:"      
}
\keyword{IO}

