% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{stringUtils}
\alias{simpleCap}
\alias{stringUtils}
\alias{substrBetween}
\alias{substrRight}
\alias{trimAll}
\alias{trimStartEnd}
\title{Utils to deal with strings and characters}
\source{
\url{http://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace-in-r}

\url{http://stackoverflow.com/questions/5992082/how-to-remove-all-whitespace-from-a-string}

\url{http://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r}

\url{http://stackoverflow.com/questions/14146362/regex-extract-string-between}

\url{http://stackoverflow.com/questions/6364783/capitalize-the-first-letter-of-both-words-in-a-two-word-string}
}
\usage{
trimStartEnd(x)

trimAll(x)

substrRight(x, n)

substrBetween(x, l, r)

simpleCap(x)
}
\arguments{
\item{x}{The string.}

\item{n}{The last n characters}

\item{l}{The regex on the left.}

\item{r}{The regex on the right.}
}
\value{
The processed string.
}
\description{
Useful function to manipulate strings and characters in \code{\link{ComMA}}.
}
\details{
\code{trimStartEnd} trims the leading and trailing whitespace in a string.

\code{trimAll} removes all whitespace from a string.

\code{substrRight} extracts the last n characters from a string x

\code{substrBetween} extracts the substring between two given regular expressions.

\code{simpleCap} capitalizes the first letter of a word string.
}
\examples{
text = "   foo bar  baz 3 "
trimStartEnd(text)
[1] "foo bar  baz 3"

text = "   foo bar  baz 3 "
trimAll(text)
[1] "foobarbaz3"

x <- "some text in a string"
substrRight(x, 6)
[1] "string"
substrRight(x, 8)
[1] "a string"

x <- "command took 0:1:34.67 (94.67s total)"
substrBetween(x, "\\\\(", "s total\\\\)")
[1] "string"

simpleCap("BACTERIA")
# for multi-words
taxaGroups <- c("BACTERIA", "FUNGI", "PROTISTS", "ANIMALIA")
sapply(taxaGroups, simpleCap)
#BACTERIA      FUNGI   PROTISTS   ANIMALIA 
#"Bacteria"    "Fungi" "Protists" "Animalia"

}
\keyword{utils}

