% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_every_n1.R
\name{sum_every_n1}
\alias{sum_every_n1}
\title{Sum Every "n"th Element, then add 1}
\usage{
sum_every_n1(x, n)
}
\arguments{
\item{x}{A numeric vector to sum over}

\item{n}{A numeric value specifying the distance between the reference index and the next index to be summed}
}
\value{
\code{sum_every_n1} returns a vector of sums of every \code{n}th element of the vector \code{x}, plus 1.
}
\description{
Sum Every "n"th Element, then add 1
}
