% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symlog}
\alias{symlog}
\title{Symmetrical Logarithm}
\usage{
symlog(x, base = exp(1), C = 0)
}
\arguments{
\item{x}{a numeric or complex vector}

\item{base}{a positive or complex number, passed to base \code{\link[=log]{log()}}. Defaults
to \emph{e}\verb{=exp(1)}.}

\item{C}{a numeric.}
}
\value{
A vector of the same length as \code{x} containing the transformed values.
Unlike \code{log}, all values \code{x} are defined for \code{symlog(x)}.
}
\description{
Symmetrical Logarithm
}
\details{
\code{y = sign(x) * log(1 + (abs(x) / 10^C))}
where \code{C} is a constant that affects how the slope of the function through
the origin. With \code{C = 0} the slope at the origin is 1.

While the output values between \code{log} and \code{symlog} are not identical,
\code{symlog} has unique behavior that can be useful in specific situations.
}
\examples{
log(-10:10)
# Returns NaN or -Inf for -10:0

symlog(-10:10)
# All values can be computed.

# The difference between symlog() and log() is max at x = 1 (delta = 0.693)
# and quickly approaches 0.
plot(1:100, symlog(1:100) - log(1:100))
}
\references{
\href{https://doi.org/10.1088/0957-0233/24/2/027001}{J Beau W Webber 2013 Meas. Sci. Technol. 24 027001}
}
