% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphTracer.R, R/Tracer.R
\name{Tracer Graph}
\alias{Tracer}
\alias{Tracer Graph}
\alias{getBurnIn}
\alias{getTraces}
\alias{gtTraces}
\alias{readMCMCLog}
\title{Tracer Graph in R}
\usage{
gtTraces(mcmc.log, burn.in, trace.id = c("posterior"), drop.1st.row = TRUE,
  line.or.point = 1, point.size = 1, point.alpha = 1, line.alpha = 0.8,
  title = "Trace", x.lab = NULL, y.lab = NULL,
  legend.title = "parameters", verbose = TRUE, ...)

readMCMCLog(file, rm.na.col = TRUE, ...)

getBurnIn(samples, burn.in = NULL, burn.in.perc = 0.1, verbose = TRUE)

getTraces(mcmc.log, burn.in = NULL, burn.in.perc = 0.1, verbose = TRUE)
}
\arguments{
\item{mcmc.log}{The data frame from MCMC log file whose 
column names are parameters and row names are the number 
of states at each sample.}

\item{burn.in}{The number of states of burn in. Default to NULL 
to use \code{burn.in.perc}. Otherwise, use \code{burn.in}.}

\item{trace.id}{A vector of name(s) of trace(s) to plot. 
Default to "posterior".}

\item{drop.1st.row}{If TRUE, also the default, then remove 
the 1st row of MCMC log.}

\item{line.or.point}{If 1, the default, then display the line only; 
if 2, then only points; if 3, then both the line and points.}

\item{point.size}{The size of points. Default to 1.}

\item{...}{Other arguments passed to \code{\link{readFile}}.}

\item{file}{The file to read/write.}

\item{rm.na.col}{If TRUE, then remove all columns with all 
missing values (NA). Default to TRUE.}

\item{burn.in.perc}{The percentage of total chain length to be burn in. 
Default to 0.1. Only used if \code{burn.in} is NULL.}

\item{...}{Other arguments passed to \code{\link{readFile}}.}

\item{mcmc.log}{The data frame from MCMC log file whose 
column names are parameters and row names are the number 
of states at each sample.}
}
\description{
A simplified implementation of Tracer \url{http://beast.bio.ed.ac.uk/tracer} in R.  
It can read MCMC log files from BEAST 1 \url{http://beast.bio.ed.ac.uk},
BEAST 2 \url{http://www.beast2.org}, or MrBayes \url{http://beast.bio.ed.ac.uk}.

It aims to make batch processing of multipe logs easier.

A simplified implementation of Tracer \url{http://beast.bio.ed.ac.uk/tracer} in R.  
It can read MCMC log files from BEAST 1 \url{http://beast.bio.ed.ac.uk},
BEAST 2 \url{http://www.beast2.org}, or MrBayes \url{http://beast.bio.ed.ac.uk}.

It aims to make batch processing of multipe logs easier.
}
\details{
\code{gtTraces} returns a \code{\link{gtable}} object to plot 
the trace given MCMC log data frame, 
whose column names are parameters and row names are the number 
of states at each sample.

\code{readMCMCLog} reads MCMC log files and return a data frame 
whose column names are parameters and row names are the number 
of states at each sample.

\code{getBurnIn} returns a list of statistics inlcuding \code{burn.in} 
and \code{step.size}, given the data frame read from \code{readMCMCLog}.
It also calculate \code{burn.in} from \code{burn.in.perc}, if it is NULL.

\code{getTraces} preprocesses the data frame returned from 
\code{readMCMCLog}.
If \code{burn.in} is NULL, then use \code{burn.in.perc}, else 
consider \code{burn.in} first.
}
\examples{
mcmc.log <- readMCMCLog("data-raw/star.beast.log")
burn.in.stats <- getBurnIn(rownames(mcmc.log))
burn.in.stats
names(mcmc.log)
gt <- gtTraces(mcmc.log, burn.in.stats$burn.in)
plot(gt)

gt <- gtTraces(mcmc.log, burn.in.stats$burn.in, trace.id=c("TreeHeight.Species", "TreeHeight.t:tree_0_0", "TreeHeight.t:tree_0_1"))

mcmc.log <- readMCMCLog("data-raw/star.beast.log")

mcmc.traces <- getTraces(mcmc.log)

mcmc.traces <- getTraces(mcmc.log)

}
\keyword{Tracer}

