% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeUtils.R
\name{TreeUtils}
\alias{TreeUtils}
\alias{annotateRAXMLTree}
\alias{renameTips}
\title{Utils to manipulate or visualise phylogenetic tree}
\usage{
annotateRAXMLTree(tree, taxa.map, anno.tree.file = "annotated-tree.nex",
  no.match = "unknown")

renameTips(tree, pattern, replacement, ...)
}
\arguments{
\item{tree}{The \pkg{\link{ape}} tree object required.}

\item{taxa.map}{The data.frame to map tree tips to traits. 
It must have at least 2 columns, where the 1st column is tree tip label.

For example,
\tabular{rrrr}{
  tip.label \tab trait1 \tab trait2 \tab ...\cr
  tip1 \tab x \tab a \tab ...\cr
  tip2 \tab x \tab b \tab ...\cr
  ... \tab y \tab a \tab ...
}}

\item{anno.tree.file}{The file path to load a tree 
that can be imported into \strong{FigTree} including annotations.
If NA, as default, to overwrite the \code{tree.file}.}

\item{pattern, replacement, ...}{Arguments of \code{\link{gsub}}.}

\item{tree.file}{The file path to load a tree generated by \strong{RAxML}.}
}
\description{
The \strong{tree} object here is \code{\link{ape}} tree.
}
\details{
\code{annotateRAXMLTree} annotates a \strong{RAxML} tree to visualise in \strong{FigTree}.

\code{renameTips} renames tree tips. It also can be used to insert annotations
using regular expression.
}
\examples{
tree <- read.tree("RAxML_bestTree.16s.txt")
tree1 <- renameTips(tree, taxa.map, "16s_RAxML.nex")

# remove "_Viridiplantae" postfix from tips
tree2 <- renameTips(tree, "_Viridiplantae", "")

# AB1_Soil.123_Metazoa:0.016 => AB1_Soil.123_Metazoa[&taxon=Metazoa]:0.016
tree2 <- renameTips(tree, "([a-zA-Z0-9\\\\_\\\\.]+)_([[a-zA-Z]+):", "\\\\1_\\\\2[\\\\&taxon=\\\\2]:")

}
\keyword{tree}

