% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsCombine.R
\name{UtilsCombine}
\alias{UtilsCombine}
\alias{combineTriMatrix}
\alias{combineTwoDF}
\alias{getTriMatrix}
\alias{mergeBy}
\alias{mergeListOfDF}
\title{Utils to combine data frames or matrices into the required data format}
\usage{
getTriMatrix(df, row.col = c("l1", "l2"), value = "corr", na.to.0 = TRUE,
  order.by = c())

combineTriMatrix(tri.m1, tri.m2)

combineTwoDF(dfm, dfm2, rm.zero = TRUE, return.df = TRUE, ...)

mergeBy(x, y, by = "row.names", rm.by = FALSE, warning.msg = TRUE, ...)

mergeListOfDF(df.list, by = "row.names", rm.by = TRUE, suffixes = c(),
  ...)
}
\arguments{
\item{df}{The data frame containing pairwised comparison result, 
such as correlations.  
For example, a data frame can be 
\tabular{rrrr}{
  l1 \tab l2 \tab corr\tab sign\cr
  16S \tab 18S \tab 0.827 \tab 0.001\cr
  16S \tab ITS \tab 0.585 \tab 0.001\cr
  18S \tab ITS \tab 0.729 \tab 0.001
}}

\item{value}{The column's values are going to fill in the output matrix.}

\item{na.to.0}{Logical, replace all NA to 0, as default.}

\item{order.by}{A vector to order matrix rows and columns. 
It must be a subset or same as their names.}

\item{tri.m1, tri.m2}{Two symmetric triangular matrices, 
which can be the output from \code{getTriMatrix}.}

\item{dfm}{A data frame or matrix.}

\item{dfm2}{The 2nd data frame or matrix whose values are into brackets.}

\item{rm.zero}{Default to TRUE to remove all " (0)".}

\item{return.df, ...}{Default to TRUE to return a data frame, otherwise a matrix.}

\item{x, y}{data frames, or objects to be coerced to one.}

\item{by, ...}{Specifications of the column(s) used for merging, 
and more arguments passed to \code{\link{merge}}.}

\item{rm.by}{Logical; if TRUE, then remove the column(s) specified
by \code{by} after merge. Default to FALSE. 
But if \code{by="row.names"}, then copy the column 'Row.names' created 
by \code{\link{merge}} into rownames before drop it.}

\item{warning.msg}{Logical; if TRUE as default, then print warning message 
if any row is dropped after merge.}

\item{df.list}{a list of data frames, or objects to be coerced to one.}

\item{suffixes}{The vector of suffixes added to distinguish colmun names 
merged by different data frames. Its length must equal to \code{df.list}.}

\item{key}{Two column names containing the pair names,
which determine the output matrix's row names and column names.
It must make row and column names same.}
}
\details{
\code{getTriMatrix} converts pairwised comparison result 
into a symmetric triangular matrix.
The pairwised comparison result is stored in a data frame.
Two of their columns must have pair names in the same order, 
which determine the matrix's row names and column names.
The 1st data frame goes to the lower triangle of the matrix, 
and the 2nd is in upper triangle.

\code{combineTriMatrix} combines two symmetric triangular matrices into one.
The 1st matrix goes to the lower triangle in the combined matrix,
and the 2nd to upper triangle.

\code{combineTwoDF} combines two data frames or matrices with a same structure in one, 
put all values in 2nd data frame into brackets.

\code{mergeBy} is an improved function of 
\code{\link{merge}} \code{by} column(s).

\code{mergeListOfDF} merges a list of data frames into one 
using \code{\link{mergeBy}}.
}
\examples{
corr.tri <- getTriMatrix(corr.df)

corr.sign.tri <- combineTriMatrix(corr.tri, sign.tri)

df <- combineTwoDF(df, df2, stringsAsFactors=FALSE)

df <- mergeBy(df, df2)

df <- mergeListOfDF(df.list, suffixes=names(df.list))
df <- mergeListOfDF(df.list, by="gene", suffixes=c("16S","18S","ITS"))

}
\keyword{utils}

