% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsSequences.R
\name{utilsSeq}
\alias{getTaxaMap}
\alias{renameFastaID}
\alias{rmDuplicateSeq}
\alias{subsetSequences}
\alias{utilsSeq}
\title{Utils to manipulate sequences or OTUs}
\usage{
subsetSequences(in.file, otus.names, out.file, regex1 = "(\\\\|[0-9]+)",
  regex2 = "", ignore.case = TRUE, max.seq = NA)

renameFastaID(fasta, regex1 = NA, regex2 = "", ignore.case = TRUE)

rmDuplicateSeq(in.file, out.file = "unique-alg.fasta")

getTaxaMap(in.file, trait.name = "taxon", trait.value = "Bacteria",
  regex1 = NA, regex2 = "", ignore.case = TRUE)
}
\arguments{
\item{in.file}{The fasta file of OTU representive sequences. 
Read by \code{\link{readFasta}}.}

\item{otus.names}{The vector of names to match sequence labels in \code{in.file}.}

\item{out.file}{The output fasta file containing extrated sequences.}

\item{regex1, regex2}{Use for \code{\link{gsub}(regex1, regex2, id(fasta))} 
to remove or replace annotation from original labels. 
Default to \code{regex1="(\\|[0-9]+)", regex2=""} in \code{subsetSequences}, 
which removes size annotation seperated by "|", 
but NA in \code{renameFastaID}, which does nothing.}

\item{ignore.case}{Refer to \code{\link{gsub}}.}

\item{max.seq}{Give the number (\code{max.seq}) of seleceted sequences, 
if extracted sequences \code{> max.seq}, then choose \code{max.seq} sequences randomly from them. 
Defaul to NA to ignore it.}

\item{fasta}{The fasta object returned by \code{\link{readFasta}}.}

\item{trait.name, trait.value}{The trait to annotate a tree. 
Its format in the string will look like '[trait.name=trait.value]'.}
}
\description{
Utils to manipulate sequences or OTUs, 
such as extract OTUs given the subset of OTU names.
They are depended on Bioconductor packages  
\url{http://www.bioconductor.org/}.
}
\details{
\code{subsetSequences} returns the subset of OTUs matching given OTU names. 
It is depended on \code{\link{ShortRead}} package. Follow the instruction 
\url{https://bioconductor.org/packages/release/bioc/html/ShortRead.html} to install.

\code{renameFastaID} renames the sequences id loaded from a fasta file 
using \code{\link{ShortRead}} package.

\code{rmDuplicateSeq} removes duplicate sequences or alginments.

\code{getTaxaMap} returns a data frame of mapping file to annotate a tree, 
such as \code{\link{annotateRAXMLTree}}.
}
\examples{
subsetSequences(in.file, otus.names, out.file)

fasta <- renameFastaID(fasta)

rmDuplicateSeq("alg.fasta", "unique-alg.fasta")

taxa.map <- getTaxaMap("16s-otus-Bacteria.fasta", trait.name="taxon", trait.value="Bacteria")

}
\keyword{util}

