% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsTaxa.R
\name{utilsTaxa}
\alias{assignTaxa}
\alias{subsetTaxaTable}
\alias{utilsTaxa}
\title{Utils to preprocess taxa table}
\usage{
subsetTaxaTable(taxa.table, taxa.group = "assigned", rank = "kingdom",
  include = TRUE, verbose = TRUE)

assignTaxa(community.matrix, taxa.table, classifier = "MEGAN",
  min.conf = 0.8, has.total = 1, col.names = c("kingdom", "phylum",
  "class", "order", "family", "genus"))
}
\arguments{
\item{taxa.table}{A data frame to contain taxonomic classifications of OTUs. 
Columns are taxonomy at the rank or lineage, rows are OTUs which need to 
match rows from community matrix. Use \code{\link{readTaxaTable}} to get it from file.}

\item{taxa.group}{The taxonomic group, the values can be 'all', 'assigned', or 
Group 'all' includes everything.
Group 'assigned' removes all uncertain classifications including 
'root', 'cellular organisms', 'No hits', 'Not assigned'. 
Alternatively, any high-ranking taxonomy in your taxonomy file 
can be used as a group or multi-groups (seperated by "|"), 
such as 'BACTERIA', 'Proteobacteria', etc. But they have to be 
in the same rank column in the file. Default to remove all 
uncertain classifications, even when group(s) assigned.}

\item{rank}{The rank column in the file to specify where to 
search for \code{taxa.group}.}

\item{include}{Define whether include or exclude given \code{taxa.group}. 
Default to TRUE.}

\item{verbose}{More details. Default to TRUE.}

\item{community.matrix}{Community matrix (OTU table), where rows are 
OTUs or individual species and columns are sites or samples. See \code{\link{ComMA}}.}

\item{classifier}{The classifier is used to generate \code{taxa.table}. 
Value is MEGAN or RDP. Default to MEGAN.}

\item{min.conf}{The confidence threshold to drop rows < \emph{min.conf}.}

\item{has.total}{If 0, then only return abudence by samples (columns) of community matrix. 
If 1, then only return toal abudence. If 2, then return abudence by samples (columns) and total. 
Default to 1.}

\item{col.names}{A vector or string of column name(s) of taxonomic ranks in the taxa table, 
which will determine the aggregated abundence matrix. They have to be full set or subset of 
c("superkingdom", "kingdom", "phylum", "class", "order", "family", "genus", "species"). 
Default to c("kingdom", "phylum", "class", "order", "family", "genus").}
}
\description{
Utils to preprocess taxa table, 
and make it easy for visualization.
}
\details{
\code{subsetTaxaTable} takes or excludes a subset of given a taxa table at given rank.

\code{assignTaxa} provides a list of taxonomic assignments with abudence 
from community matrix. The function is iterated through \code{col.names}, 
and \code{\link{aggregate}}s abudence into taxonomy based on the rank in 
\code{col.names}. All sequences either classified as 
"root|cellular organisms|No hits|Not assigned|unclassified sequences"
from BLAST + MEGAN, or confidence < \emph{min.conf} threshold from RDP, 
are changed to "unclassified", which will be moved to the last row.
}
\examples{
tt.sub <- subsetTaxaTable(tt.megan, taxa.group="Proteobacteria", rank="phylum")
tt.sub <- subsetTaxaTable(tt.megan, taxa.group="Cnidaria|Brachiopoda|Echinodermata|Porifera", rank="phylum", include=FALSE)

ta.list <- assignTaxa(community.matrix, tt.megan)
ta.list <- assignTaxa(community.matrix, tt.rdp, col.names="kingdom", classifier="RDP")

}
\keyword{utils}

