% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMethylSites.R
\name{annotateMethylSites}
\alias{annotateMethylSites}
\title{Annotate methylation sites}
\usage{
annotateMethylSites(methyl_df, meta_df, location)
}
\arguments{
\item{methyl_df}{A dataframe with a column listing every methylated site of interest, for example adenines within GATC sites.}

\item{meta_df}{A dataframe containing columns Type (feature type e.g. Promoter), Site (name of site e.g. thrLp), Left, and Right (for start and end of each feature).}

\item{location}{Column within methyl_df containing the genomic position of the methylation sites}
}
\value{
Dataframe methyl_df with columns added for genomic features present in each methylation site.
}
\description{
Takes a dataframe with a numeric vector for methylated sites of interest (for example, all adenines within GATC sites) and a dataframe with the genomic locations of features of interest (like Promoters, Sigma factor binding sites, genes, etc.) and determines which features (can be more than one) are present at each methylation site.
}
\examples{
df <- read.table(file="./data/WT_6mA_Mg.txt", header = T, sep = '\t')
meta <- read.table('./all_site_annotations.txt', header = T, sep = '\t')
df_annotated <- annotateMethylSites(df, meta, location='Position')
}
