% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMethylSites.R,
%   R/annotateMethylSites_exp.R
\name{annotateMethylSites}
\alias{annotateMethylSites}
\title{Annotate Methylated Sites}
\usage{
annotateMethylSites(methyl_df, meta_df, location)

annotateMethylSites(methyl_df, meta_df, location)
}
\arguments{
\item{methyl_df}{A data frame containing information on methylated sites, with columns "Chr", "Start", and "End" indicating the chromosome and genomic coordinates of each site.}

\item{meta_df}{A data frame containing metadata on genomic features, with columns "Chr", "Left", "Right", "Type", and "Site". "Left" and "Right" specify the genomic coordinates of the feature, "Type" specifies the type of feature (e.g. exon, intron, promoter), and "Site" specifies a unique identifier for the feature.}

\item{location}{A character string indicating which column in methyl_df should be used to determine the location of each methylated site (e.g. "Start" or "End").}
}
\value{
A data frame containing the same columns as methyl_df, with additional columns added for each feature type present in meta_df. Each row represents a methylated site, and the value in each feature column indicates the unique identifier(s) of the feature(s) that the site overlaps with. If a site does not overlap with any feature, the "No_Feature" column will be set to 1.

A data frame containing the same columns as methyl_df, with additional columns added for each feature type present in meta_df. Each row represents a methylated site, and the value in each feature column indicates the unique identifier(s) of the feature(s) that the site overlaps with. If a site does not overlap with any feature, the "No_Feature" column will be set to 1.
}
\description{
This function takes as input a data frame of methylated sites and a metadata data frame, and annotates each methylated site with its corresponding feature(s) based on the given location. If a site does not overlap with any feature in the metadata, it is labeled as "No_Feature".

This function takes as input a data frame of methylated sites and a metadata data frame, and annotates each methylated site with its corresponding feature(s) based on the given location. If a site does not overlap with any feature in the metadata, it is labeled as "No_Feature".
}
\examples{
df <- methyl_df
meta <- genome_sites
df_annotated <- annotateMethylSites(df, meta, location='Position')

# Load example data
data(methyl_df)
data(meta_df)

# Annotate methylated sites with genomic features
annotated_df <- annotateMethylSites(methyl_df, meta_df, "Start")

}
\seealso{
\code{\link{findOverlaps}}, \code{\link{GenomicRanges-class}}

\code{\link{GenomicRanges-class}}
}
