% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateTSS.R
\name{annotateTSS}
\alias{annotateTSS}
\title{Title}
\usage{
annotateTSS(methyl_df, meta_df, location, size, long = TRUE)
}
\arguments{
\item{methyl_df}{A dataframe.}

\item{meta_df}{A dataframe.}

\item{location}{A string.}

\item{size}{An integer.}

\item{long}{A boolean.}
}
\value{
A dataframe.
}
\description{
Title
}
\examples{
data(WT_average_annotated)
WT_all_TSS <- annotateTSS(WT_average_annotated, genome_sites, location = 'Position', size = 500)
}
