% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{buildMicrobeMethyl}
\alias{buildMicrobeMethyl}
\title{Build MicrobeMethyl object from bedMethyl file}
\usage{
buildMicrobeMethyl(bedMethyl_file, sample_name, metadata)
}
\arguments{
\item{bedMethyl_file}{Path to the bedMethyl file (a tab-separated file).}

\item{sample_name}{A character string representing the name of the sample.}

\item{metadata}{A list containing metadata for the sample, such as
experimental conditions and any other relevant information (optional).}
}
\value{
A MicrobeMethyl object.
}
\description{
Reads a bedMethyl file (as defined by the ENCODE consortium) and creates
a MicrobeMethyl object containing methylation data and sample metadata for a
single sample.
}
