% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityAnalysis.R
\name{classifyByTaxaComp}
\alias{classifyByTaxaComp}
\title{Classify sample attribute by taxonomic composition}
\usage{
classifyByTaxaComp(data, attr.data, group.id, percent = 0.01, alpha = 1,
  family = "multinomial", nlambda = 500, ...)
}
\arguments{
\item{data, attr.data, group.id, percent}{Refer to \code{\link{enterotype}}.}

\item{alpha, family, nlambda}{Refer to \code{\link{cv.glmnet}}.}
}
\details{
Build a classification model using lasso \code{\link{cv.glmnet}},
to predict the certain type of samples (e.g. land types) 
given relative abundance of taxonomic compositions (e.g. families) in the samples, 
and also to find which taxonomic compositions are important to express this model.

This can be used after
}
\examples{
cvfit <- classifyByTaxaComp(abun.dist.matrix, attr.data=env, group.id="land.use")
}
\keyword{classification}

