% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoMMA-package.R
\docType{package}
\name{CoMMA-package}
\alias{CoMMA}
\alias{CoMMA-package}
\title{CoMMA: Comparison of Microbial Methylated Adenines}
\description{
Comparative adenine methylomics between bacterial samples. This package can take methylated adenine calls from nanopore sequencing as input, then characterize methylation throughout the genome, identify differentially methylated sites between samples, and compare full adenine methylomes between samples.
}
\author{
\strong{Maintainer}: Carl Stone \email{carl.j.stone@vanderbilt.edu} (\href{https://orcid.org/0000-0003-0232-5223}{ORCID})

}
\keyword{internal}
