% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComMA.R
\docType{package}
\name{ComMA-package}
\alias{ComMA}
\alias{ComMA-package}
\title{ComMA R documentation}
\source{
\url{https://github.com/walterxie/ComMA}

Install all depdent packages: \code{\link{installAllPackages()}}
}
\description{
This package is a collection of tools for community matrix analyse, 
especially for eDNA data, but it also can be used for traditional species survey data.
The package not only wraps bioinformatic, ecological methods and figure plotting  
into one-line functions, but also create utility functions to procced data easily. 
All functions are available by typing \code{help(package="ComMA")} in R.
}
\details{
\strong{Community matrix} in \pkg{ComMA} is a community data from file as a matrix,
where rows are OTUs (therefore the matrix is called \emph{OTU table}) or individual species 
and columns are sites or samples. 
Matrix elements are abundance data or proportion (e.g. counts, percentage). For example,
\tabular{rrrr}{
  OTU_id \tab plot01 \tab plot02\tab ...\cr
  OTU_1 \tab 1 \tab 0 \tab ...\cr
  OTU_2 \tab 100 \tab 200 \tab ...\cr
  OTU_3 \tab 56 \tab 3 \tab ...
}

Variable \strong{t.community.matrix} is a transposed matrix from community matrix 
we defined here, where columns are OTUs or individual species and rows are sites or samples. 
It is also the abundances argument in \pkg{vegetarian} \code{\link{d}}.

The \strong{taxaPaths} from taxonomic path file is a taxonomic matrix, 
where rows are OTUs or individual species (they have to be the subset of rows 
in the community matrix from \code{getCommunityMatrix}), 
and columns are taxonomic ranks, such as c("superkingdom", "kingdom", "phylum", 
"class", "order", "family", "genus", "species") inlcuding full taxonomic path.

The enviornmental data \strong{env} is also called as the meta data of samples, 
where rows are sites or samples, and columns are the measurement, such as 
elevation, tempurature, soil chemistry, forest type, etc.

The \strong{phyloTree} is a rooted tree of \code{\link{phylo}} object, 
which is created from \pkg{ape} \code{\link{read.tree}}.
}

