% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityComparison.R
\name{CommunityComparison}
\alias{CommunityComparison}
\alias{getDissimilarityList}
\alias{mantelComparison}
\alias{plotProcrustes}
\alias{procrustesComparison}
\alias{sublistByPairs}
\title{Compare community structure}
\usage{
getDissimilarityList(cm.list, metric = "jaccard", ...)

mantelComparison(dist.list, method = "pearson", permutations = 999)

procrustesComparison(dist.list, scale = TRUE, symmetric = TRUE,
  permutations = 999)

sublistByPairs(m.p.list, pairs, subset.pairs)

plotProcrustes(proc.list, attr.df, colour.id = "Elevation",
  proc.list.pairs = list(), title.list = list(), plot.title.size = 8,
  colours = c("blue", "orange"), limits = NULL,
  legend.title = "Elevation (m)", legend.key.width = unit(0.65, "cm"))
}
\arguments{
\item{cm.list}{A list of community matrices.}

\item{metric, ...}{The distance metric, default to "jaccard",
and other parametes passed to \code{\link{getDissimilarity}}.}

\item{dist.list}{A list of dissimilarity in \code{\link{dist}} 
between pairwised community matrices, 
which can be calculated by \code{\link{getDissimilarityList}}.}

\item{method, permutations}{Parameters used in \code{\link{mantel}} in \pkg{vegan}.}

\item{scale, symmetric, permutations}{Parameters used in 
\code{\link{procrustes}} and \code{\link{protest}}.}

\item{m.p.list}{Either \code{m.list} of the output from \code{mantelComparison}, 
or \code{proc.list} or \code{prot.list} from \code{procrustesComparison}.}

\item{pairs, subset.pairs}{The list of the list of two pairs of community names, 
see example. 
The string names have to be exaclty same in order to find matches, 
but the order of the pair does not matter the 1st or 2nd come first.}

\item{proc.list}{The list of \code{\link{procrustes}} results.}

\item{attr.df, colour.id}{A data frame of meta data to define 
how the plot is coloured by \code{colour.id}. Default to "Elevation".}

\item{title.list, proc.list.pairs}{The list of strings to make titles. 
Use \code{title.list} first, and then \code{proc.list.pairs} if \code{title.list} is empty, 
and then \code{names(proc.list)} if both are empty.}

\item{colours, legend.title, legend.key.width}{The properties of colour legend}
}
\description{
Multivariate patterns of sample similarity were compared between pairwised community matrices 
using Procrustes and Mantel tests.
}
\details{
\code{getDissimilarityList} calculates a list of similarity/dissimilarity 
given a list of community matrices and the distance metric.

\code{mantelComparison} makes Mantel tests for comparisons 
between the overall community structure.

\code{procrustesComparison} makes Procrustes comparisons 
between the overall community structure.

\code{sublistByPairs} takes a subset result from comparisons 
given the subset of pairwised communities.

\code{plotProcrustes} plots Procrustes correlations between pairwised communities.
}
\examples{
jaccard.dist <- getDissimilarityList(cm.list, metric="jaccard")
jaccard.dist$dist.list

mantel <- mantelComparison(jaccard.dist$dist.list)
mantel.tri <- getTriMatrix(mantel$m.df) 

procrustes <- procrustesComparison(jaccard.dist$dist.list) 
prot.tri <- getTriMatrix(procrustes$prot.df, order.by=order.by) 

# Mantel test (lower triangle) and Procrustes test (upper triangle)
corrs <- combineTriMatrix(mantel.tri, prot.tri)

sub.proc <- sublistByPairs(procrustes$proc.list, procrustes$pairs, subset.pairs=list(list("16S bacteria","18S protists"),list("18S animals","COI−300 animals")) )
sub.proc$sub.list

plotProcrustes(procrustes$proc.list)

}
\keyword{comparison}

