% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityComparison.R
\name{CommunityComparison}
\alias{CommunityComparison}
\alias{getCorrTriMatrix}
\alias{getDissimilarityList}
\alias{mantelComparison}
\alias{procrustesComparison}
\title{Compare community structure}
\usage{
getDissimilarityList(cm.list, metric = "jaccard", ...)

mantelComparison(dist.list, method = "pearson", permutations = 999)

procrustesComparison(dist.list, scale = TRUE, symmetric = TRUE,
  permutations = 999)

getCorrTriMatrix(mantel.df, prot.df)
}
\arguments{
\item{cm.list}{A list of community matrices.}

\item{metric, ...}{The distance metric, default to "jaccard",
and other parametes passed to \code{\link{getDissimilarity}}.}

\item{dist.list}{A list of dissimilarity in \code{\link{dist}} 
between pairwised community matrices, 
which can be calculated by \code{\link{getDissimilarityList}}.}

\item{method, permutations}{Parameters used in \code{\link{mantel}} in \pkg{vegan}.}

\item{scale, symmetric, permutations}{Parameters used in 
\code{\link{procrustes}} and \code{\link{protest}}.}

\item{mantel.corr.list, prot.corr.list}{The list of correlations from
Mantel test and Procrustes test.}
}
\description{
Multivariate patterns of sample similarity were compared between pairwised community matrices 
using Procrustes and Mantel tests.
}
\details{
\code{getDissimilarityList} calculates a list of similarity/dissimilarity 
given a list of community matrices and the distance metric.

\code{mantelComparison} makes Mantel tests for comparisons 
between the overall community structure.

\code{procrustesComparison} makes Procrustes comparisons 
between the overall community structure.

\code{getCorrTriMatrix} combines Mantel test (lower triangle) 
and Procrustes test (upper triangle) correlations into one matrix 
for comparisons.
}
\examples{
jaccard.dist <- getDissimilarityList(cm.list, metric="jaccard")
jaccard.dist$dist.list

mantel <- mantelComparison(jaccard.dist$dist.list)

procrustes <- procrustesComparison(jaccard.dist$dist.list)

corr.tri.m <- getCorrTriMatrix(mantel$m.df, procrustes$prot.df)

}
\keyword{comparison}

