% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunityComparison.R
\name{comparison}
\alias{comparison}
\alias{getDissimilarityList}
\alias{mantelComparison}
\title{Compare community structure}
\usage{
getDissimilarityList(cm.list, metric = "jaccard", ...)

mantelComparison(dist.list, method = "pearson", permutations = 999)
}
\arguments{
\item{cm.list}{A list of community matrices.}

\item{metric, ...}{The distance metric, default to "jaccard",
and other parametes passed to \code{\link{getDissimilarity}}.}

\item{dist.list}{A list of dissimilarity in \code{\link{dist}} 
between pairwised community matrices, 
which can be calculated by \code{\link{getDissimilarityList}}.}

\item{method, permutations}{Parameters used in \code{\link{mantel}} in \pkg{vegan}.}
}
\description{
Multivariate patterns of sample similarity were compared between pairwised community matrices 
using Procrustes and Mantel tests.
}
\details{
\code{getDissimilarityList} calculates a list of similarity/dissimilarity 
given a list of community matrices and the distance metric.

Mantel tests for comparisons between the overall community structure.
}
\examples{
dist.list <- getDissimilarityList(cm.list, metric="jaccard")

corr.sign <- mantelComparison(dist.list)

}
\keyword{classification}
\keyword{comparison}

