% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxaCountOTUsReads.R
\name{countOTUsReads}
\alias{countOTUsReads}
\alias{getCountSums}
\alias{plotTaxonomy}
\alias{sumReadsOTUs}
\title{Taxonomy analyses for multiple data sets}
\usage{
getCountSums(..., input.list = FALSE, taxa.rank = "phylum",
  group.rank = "kingdom")

plotTaxonomy(all.counts.sums, taxa.ref = ComMA::taxa.ref.PLOSONE.2015,
  taxa.rank = "phylum", group.rank = "kingdom", render.rdp = FALSE,
  gene.levels = c("16S", "18S", "26S", "ITS", "COI-300", "COI-650"),
  group.levels = c("ARCHAEA", "BACTERIA", "EUKARYOTA", "PROTOZOA",
  "CHROMISTA", "FUNGI", "PLANTAE", "ANIMALIA", "Unknown"),
  exclude.singletons = 1, x.lab = "Phylum (or higher-level taxon)",
  y.lab = "Number of sequences or OTUs", legend.title = NULL,
  palette = NULL, title = "", title.size = 10, verbose = TRUE)

sumReadsOTUs(cm.taxa.list, taxa.ref = ComMA::taxa.ref.PLOSONE.2015,
  taxa.rank = "phylum", group.rank = "kingdom", col.ranks = c("kingdom",
  "phylum", "class", "order"), gene.levels = c("16S", "18S", "26S", "ITS",
  "COI-300", "COI-650"), group.levels = c("ARCHAEA", "BACTERIA", "EUKARYOTA",
  "PROTOZOA", "CHROMISTA", "FUNGI", "PLANTAE", "ANIMALIA", "Unknown"),
  x.lab = "Phylum (or higher-level taxon)",
  y.lab = "Number of sequences or OTUs", fig.folder = "./figures",
  table.folder = "./outputs", pdf.width = 260, pdf.height = 200,
  units = "mm")
}
\arguments{
\item{...}{Input of a list of community matrices, or comma separated multi-inputs.}

\item{input.list}{Default to TRUE to unwrap list(...) to 
get the actual list if the input is a list of cm.}

\item{taxa.rank}{The rank to display in the axis for counts of OTUs and reads.}

\item{group.rank}{The rank to group \code{taxa.rank} to colour the figure.}

\item{all.counts.sums}{Teh output from \code{getCountSums}.}

\item{taxa.ref}{The taxonomic reference data set to order taxonomic names nicely.
But it removes the taxonomy not exsiting in the reference.
Default to use a taxonomy reference data set \code{taxa.ref.PLOSONE.2015} in the package.
Set \code{taxa.ref=""}, where length(taxa.ref) > 1, to plot all taxonomy.}

\item{render.rdp}{Render the taxonomy from RDP when using \code{taxa.ref.PLOSONE.2015}, 
the default is FALSE.}

\item{gene.levels}{The level to order 'gene' column.}

\item{group.levels}{The level to order 'gene' column.}

\item{exclude.singletons}{If 0, then do not plot the number of 
sequences and OTUs excluding singleton OTUs. 
If 1, as default, just do not plot sequences excluding singleton. 
If 2, then plot all (including/excluding singleton).}

\item{x.lab, y.lab}{Label for x or y axis.}

\item{palette}{The detail in \code{\link{ggPlot}}.}

\item{cm.taxa.list}{A list of \code{cm.taxa} merged by \code{\link{mergeCMTaxa}}. 
Require proper names of the list.}

\item{col.ranks}{A vector or string of column name(s) of taxonomic ranks in the taxa table, 
detail to see \code{\link{mergeCMTaxa}}. 
Default to c("kingdom", "phylum", "class", "order").}

\item{fig.folder}{Folder to contain the figure in PDF from \code{plotTaxonomy}.}

\item{table.folder}{Folder to save the result from \code{getCountSums}.}

\item{pdf.width, pdf.height, units}{Parameters in \code{\link{ggsave}} to adjust figure.}
}
\description{
Taxonomic summary of OTUs and reads for multiple data sets, 
and bar chart for visualization.
"count.1" column is number of OTUs including singletons, 
"sum.1" is number of reads including singletons, 
"count.2" is OTUs excluding singletons, 
"sum.2" is reads excluding singletons.
}
\details{
\code{getCountSums} return a list of counts of OTUs and reads given a list of
data sets.

\code{plotTaxonomy} makes a plot of OTU and sequence counts by taxonomic group 
sumarised by \code{getCountSums}.

Diamonds represent the number of sequences, 
open circles the number of OTUs including singleton OTUs, 
and filled circles the number of OTUs excluding singleton OTUs.

\code{sumReadsOTUs} is a high-level function to use \code{getCountSums} to sumarise counts,
and call \code{plotTaxonomy} to plot figure.
}
\examples{
all.counts.sums <- ComMA::getCountSums(cm.taxa.list, input.list=T, group.rank="kingdom", taxa.rank="phylum")

plotTaxonomy(all.counts.sums, taxa.ref=taxa.ref)


}
\keyword{taxonomy}

