% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiverRare.R
\name{diverrare}
\alias{diverrare}
\alias{getDiverRare}
\alias{getSubsamplesForDiverRare}
\title{Rarefaction curves using Jost diversities}
\usage{
getSubsamplesForDiverRare(cm, sample.size, replicates = 10,
  levels = rep(c("gamma", "alpha", "beta"), 3), qs = rep(0:2, each = 3),
  is.transposed = FALSE, progressBar = TRUE, verbose = TRUE)

getDiverRare(cm, sample.sizes = 10, replicates = 10,
  levels = rep(c("gamma", "alpha", "beta"), 3), qs = rep(0:2, each = 3),
  progressBar = TRUE, min.sample.abundance = NA,
  non.rarefied.subsamples = TRUE, verbose = TRUE)
}
\arguments{
\item{cm}{The community matrix.}

\item{sample.size}{An interger number representing subsample size for 
rarefying community using \code{\link{rrarefy}}.
It must be smaller than the minimun sample abundance, 
which can be calculated by \code{\link{summaryCM.Vector}}.}

\item{replicates}{The number of replicates to repeat subsampling.}

\item{levels}{The levels of Jost diversities in \code{\link{d}}.
Use \code{paste(levels, qs, sep="")} to get the actual measurements.}

\item{qs}{The order of Jost diversities in \code{\link{d}}.}

\item{is.transposed}{If TRUE, then \code{cm} is already
the transposed matrix of \code{\link{vegdist}} input. 
Default to FASLE.}

\item{progressBar}{Whether to print progress bar, default to TRUE.}

\item{sample.sizes}{The vector of subsample sizes for 
rarefying community using \code{\link{rrarefy}}.
It can be also an integer to define the length of vector, as default to 10, 
which allows to automatically create the vector of subsample sizes ranged by 
\code{min.sample.abundance} and \code{max.sample.abundance} of \code{cm}}

\item{min.sample.abundance}{If \code{min.sample.abundance} of \code{cm} 
is smaller than this threshold, then return NULL.}

\item{non.rarefied.subsamples}{If TRUE, as default, the subsamples will include
non rarefied ones, whose \code{sample.size} > \code{min.sample.abundance}.
\code{\link{rrarefy}} will also create warnings.}
}
\value{
\code{getSubsamplesForDiverRare} returns a list of results: 
\code{mean} is a named vector of means of each diversity measurement, 
\code{sterr} is the standard error of the mean, 
\code{subsamples} is a matrix of all subsampled points to calculate means and errors, 
where the columns are diversity measurements, rows are replicates.
}
\description{
Create rarefaction curves using Jost diversities.
}
\details{
\code{getSubsamplesForDiverRare} calculates subsamples for 
the rarefaction using Jost diversities, given a community matrix.

\code{getDiverRare} returns a list of subsamples calculated 
from \code{getSubsamplesForDiverRare} given a community matrix.
Note: this community matrix must not be transposed.
}
\examples{
diver.rare <- getSubsamplesForDiverRare(cm, 2000, 5)

diver.rare <- getDiverRare(cm, replicates=5)

}
\keyword{rarefaction}

