% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motif_sites.R
\name{.loadGenomeSequences}
\alias{.loadGenomeSequences}
\title{Load genome sequences from a FASTA path or BSgenome object}
\usage{
.loadGenomeSequences(genome)
}
\value{
A \code{DNAStringSet} with one element per chromosome.
}
\description{
Load genome sequences from a FASTA path or BSgenome object
}
\keyword{internal}
