% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUtils.R
\name{EUtils}
\alias{EUtils}
\alias{parseTSeqSet}
\alias{parseTaxaSet}
\title{EUtils NCBI}
\usage{
parseTaxaSet(xml)

parseTSeqSet(xml)
}
\arguments{
\item{xml}{The XML result of \code{\link{efetch}}.}
}
\description{
Utils to process data downloaded from NCBI.
}
\details{
\code{parseTaxaSet} parses the result XML (DOCTYPE is TaxaSet) 
of \code{\link{efetch}} from taxonomy database into a data.frame, 
which inlcudes "TaxId", "ScientificName", "Rank", "Lineage", "Division", 
and the format of taxa.table from "kingdom" to "genus".

<TaxaSet>
<Taxon>
  <TaxId>123685</TaxId>
  <ScientificName>Oryzias minutillus</ScientificName>
  <ParentTaxId>8089</ParentTaxId>
  <Rank>species</Rank>
  <Division>Vertebrates</Division>
  <Lineage>cellular organisms; Eukaryota; Opisthokonta; Metazoa; ...</Lineage>
  <LineageEx>
   <Taxon>
    <TaxId>131567</TaxId>
    <ScientificName>cellular organisms</ScientificName>
    <Rank>no rank</Rank>
   </Taxon>
   ...  
  </LineageEx>
</Taxon>
...
</TaxaSet>

\code{parseTSeqSet} parses the result XML (DOCTYPE is TSeqSet) 
of \code{\link{efetch}} from nuccore database into a data.frame, 
which inlcudes "TaxId", "ScientificName", "ACCESSION", "Lineage", "sequence".

<TSeqSet>
  <TSeq>
    <TSeq_seqtype value="nucleotide"/>
    <TSeq_gi>1079489517</TSeq_gi>
    <TSeq_accver>NC_031445.1</TSeq_accver>
    <TSeq_sid>gnl|NCBI_GENOMES|60824</TSeq_sid>
    <TSeq_taxid>126358</TSeq_taxid>
    <TSeq_orgname>Abeliophyllum distichum</TSeq_orgname>
    <TSeq_defline>Abeliophyllum distichum chloroplast, complete genome</TSeq_defline>
    <TSeq_length>155982</TSeq_length>
    <TSeq_sequence>CATTTTAGTTATGGGC...GCTGT</TSeq_sequence>
  </TSeq>
</TSeqSet>
}
\examples{
library("reutils")
taxa <- efetch(c("123685", "8089", "8088"), "taxonomy")
taxa.df <- parseTaxaSet(taxa$content)

seqset <- efetch("NC_031445.1", "nuccore", "fasta")
seqset.df <- parseTSeqSet(seqset$content)

}
\keyword{eutils}

