% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUtils.R
\name{EUtils}
\alias{EUtils}
\alias{parseTSeqSet}
\alias{parseTaxaSet}
\alias{seqSet2Fasta}
\title{EUtils NCBI}
\usage{
parseTaxaSet(xml)

parseTSeqSet(xml)

seqSet2Fasta(seqset.uid, by = 5, folder = ".", file.prefix = "Refseq",
  file.extension = "fasta", seq.label = c("accver", ", ", "orgname",
  ", chloroplast"), sleep = 30)
}
\arguments{
\item{xml}{The XML result of \code{\link{efetch}}.}

\item{seqset.uid}{The vector of \code{uid} for \code{\link{efetch}} to download sequence,
where db = "nuccore", rettype = "fasta". 
The \code{uid} will either look like NC_016668.1 ("accver") or KM462867 ("INSDC").}

\item{by}{Split \code{seqset.uid} into subgroups by the given number, defaul to 20.}

\item{folder, file.prefix, file.extension}{Determine the file name.}

\item{seq.label}{The vector of string to determine how to label the sequence. 
If the element is one of the column name data.frame from \code{parseTSeqSet}, 
then the label of that position will be the value of that column. 
The available columns are "seqtype", "gi", "accver", "taxid", "orgname", "defline", "length".}

\item{sleep}{Please set enough sleep time. Default 30 seconds.}
}
\description{
Utils to process data downloaded from NCBI.
}
\details{
\code{parseTaxaSet} parses the result XML (DOCTYPE is TaxaSet) 
of \code{\link{efetch}} from taxonomy database into a data.frame, 
which inlcudes "TaxId", "ScientificName", "Rank", "Lineage", "Division", 
and the format of taxa.table from "kingdom" to "genus".

<TaxaSet>
<Taxon>
  <TaxId>123685</TaxId>
  <ScientificName>Oryzias minutillus</ScientificName>
  <ParentTaxId>8089</ParentTaxId>
  <Rank>species</Rank>
  <Division>Vertebrates</Division>
  <Lineage>cellular organisms; Eukaryota; Opisthokonta; Metazoa; ...</Lineage>
  <LineageEx>
   <Taxon>
    <TaxId>131567</TaxId>
    <ScientificName>cellular organisms</ScientificName>
    <Rank>no rank</Rank>
   </Taxon>
   ...  
  </LineageEx>
</Taxon>
...
</TaxaSet>

\code{parseTSeqSet} parses the result XML (DOCTYPE is TSeqSet) 
of \code{\link{efetch}} from nuccore database into a data.frame, 
which inlcudes "TaxId", "ScientificName", "ACCESSION", "Lineage", "sequence".

<TSeqSet>
  <TSeq>
    <TSeq_seqtype value="nucleotide"/>
    <TSeq_gi>1079489517</TSeq_gi>
    <TSeq_accver>NC_031445.1</TSeq_accver>
    <TSeq_sid>gnl|NCBI_GENOMES|60824</TSeq_sid>
    <TSeq_taxid>126358</TSeq_taxid>
    <TSeq_orgname>Abeliophyllum distichum</TSeq_orgname>
    <TSeq_defline>Abeliophyllum distichum chloroplast, complete genome</TSeq_defline>
    <TSeq_length>155982</TSeq_length>
    <TSeq_sequence>CATTTTAGTTATGGGC...GCTGT</TSeq_sequence>
  </TSeq>
</TSeqSet>

\code{seqSet2Fasta} downloads reference sequences given their \code{uid} 
using \code{\link{efetch}}. \code{parseTSeqSet} is used to parse the 
\code{\link{efetch}} result.
}
\examples{
library("reutils")
taxa <- efetch(c("123685", "8089", "8088"), "taxonomy")
taxa.df <- parseTaxaSet(taxa$content)

seqset <- efetch("NC_031445.1", "nuccore", "fasta")
seqset.df <- parseTSeqSet(seqset$content)

seqSet2Fasta(c("NC_031445.1", "NC_026892.1"), seq.label=c("accver", ", ", "orgname", ", chloroplast")) 
 
}
\keyword{eutils}

