% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{getBetaValues,microbeMethyl-method}
\alias{getBetaValues,microbeMethyl-method}
\title{getBetaValues for microbeMethyl objects}
\usage{
\S4method{getBetaValues}{microbeMethyl}(object)
}
\arguments{
\item{object}{A microbeMethyl object}
}
\value{
A numeric vector of beta values for the given microbeMethyl object
}
\description{
Retrieve the beta values from a microbeMethyl object.
}
