% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{getBetaValues,MicrobeMethylExperiment-method}
\alias{getBetaValues,MicrobeMethylExperiment-method}
\title{getBetaValues for MicrobeMethylExperiment objects}
\usage{
\S4method{getBetaValues}{MicrobeMethylExperiment}(object)
}
\arguments{
\item{object}{A MicrobeMethylExperiment object}
}
\value{
A list of numeric vectors, where each vector contains the beta values
for a sample in the MicrobeMethylExperiment object
}
\description{
Retrieve the beta values for all samples in a
MicrobeMethylExperiment object.
}
