% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{getBetaValues,microbeMethylExperiment-method}
\alias{getBetaValues,microbeMethylExperiment-method}
\title{getBetaValues for microbeMethylExperiment objects}
\usage{
\S4method{getBetaValues}{microbeMethylExperiment}(object)
}
\arguments{
\item{object}{A microbeMethylExperiment object}
}
\value{
A list of numeric vectors, where each vector contains the beta values
for a sample in the microbeMethylExperiment object
}
\description{
Retrieve the beta values for all samples in a
microbeMethylExperiment object.
}
