% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{getBetaValues}
\alias{getBetaValues}
\title{Get beta values from MicrobeMethyl or MicrobeMethylExperiment objects}
\usage{
getBetaValues(object)
}
\arguments{
\item{object}{An object of class MicrobeMethyl or MicrobeMethylExperiment}
}
\value{
A vector of beta values for a MicrobeMethyl object, or a list of
beta value vectors for a MicrobeMethylExperiment object
}
\description{
This generic function retrieves the beta values from MicrobeMethyl or
MicrobeMethylExperiment objects.
}
