% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\name{getBetaValues}
\alias{getBetaValues}
\title{Get beta values from microbeMethyl or microbeMethylExperiment objects}
\usage{
getBetaValues(object)
}
\arguments{
\item{object}{An object of class microbeMethyl or microbeMethylExperiment}
}
\value{
A vector of beta values for a microbeMethyl object, or a list of
beta value vectors for a microbeMethylExperiment object
}
\description{
This generic function retrieves the beta values from microbeMethyl or
microbeMethylExperiment objects.
}
