% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{ggBarYAcrossX}
\alias{ggBarYAcrossX}
\title{The bar chart shows the number of OTUs in the bar across the number of samples in the value of x-axis, 
which is extended from \code{\link{ggBarChart}}.}
\usage{
ggBarYAcrossX(community.matrix,
  title = "The number of OTUs/reads across the number of samples",
  x.lab = "Number of samples crossed", y.lab = "Number of OTUs/reads",
  y.trans = "log", auto.scale.y = TRUE, x.scale = "discrete",
  x.interval = 1, x.text.angle = 0, legend.title = "", ...)
}
\arguments{
\item{community.matrix}{Community matrix (OTU table), where rows are 
OTUs or individual species and columns are sites or samples. See \code{\link{ComMA}}.}

\item{title}{Graph title}

\item{x.lab, y.lab}{The label of x-axis or y-axis, such as plot names.}

\item{...}{More from \code{\link{ggBarChart}}.}

\item{low, }{high Refer to \pkg{ggplot2} \code{\link{scale_fill_gradient}}. 
Default to low="white", high="steelblue".}

\item{autoWidth}{If TRUE, then use number of bars and legend columns 
to estimate pdf width automatically. Default to TRUE.}
}
\description{
The bar chart shows the number of OTUs in the bar across the number of samples in the value of x-axis, 
which is extended from \code{\link{ggBarChart}}.
}
\examples{
 
community.matrix <- getCommunityMatrix("16S", isPlot=TRUE, minAbund=1)
bar.yx <- ComMA::ggBarYAcrossX(community.matrix)
}
\keyword{graph}

