% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{ggNMDSPlot}
\alias{ggNMDSPlot}
\title{Use \code{\link{metaMDS}} in \code{\link{vegan}}
to create Nonmetric Multidimensional Scaling (NMDS) plot .}
\usage{
ggNMDSPlot(comm, attr.df, coloured.by, shaped.by, linked.by,
  distance = "bray", add.text = FALSE, ..., title = "MDS",
  verbose = TRUE)
}
\arguments{
\item{comm}{Community data for NMDS plot. 
It is dissimilarities either as a \code{\link{dist}} structure 
or as a community data defined by \code{\link{vegan}} which is 
a transposed matrix of community matrix in \code{\link{ComMA}}.
Use \code{\link{transposeCM}} to rotate the data frame.
Detail to \code{\link{metaMDS}}.}

\item{attr.df}{A data frame to define how NMDS plot is coloured, shaped, or linked.
where its row names have to match row names of community data \code{comm}.}

\item{distance}{Dissimilarity index used in \code{\link{vegdist}}.}
}
\description{
Use \code{\link{metaMDS}} in \code{\link{vegan}}
to create Nonmetric Multidimensional Scaling (NMDS) plot .
}
\examples{

nmds.plot <- ggNMDSPlot(comm, attr.df, coloured.by="", shaped.by="", linked.by="")
}
\keyword{graph}

