% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{ggPercentageBarChart}
\alias{ggPercentageBarChart}
\title{Percentage bar chart coloured by groups, which is extended from \code{\link{ggBarChart}}.}
\usage{
ggPercentageBarChart(df.to.melt, melt.id, title = "Percentage Bar Chart",
  x.lab = "", y.lab = "", palette = NULL, x.text.angle = 90,
  autoWidth = TRUE, ...)
}
\arguments{
\item{df.to.melt}{A data frame required to \code{\link{melt}} before making a percent bar chart. 
For example,
\tabular{rrrrr}{
  Phyla \tab 16s \tab 18s \tab ITS \tab TaxaGroup\cr
  Actinobacteria \tab 958 \tab 1 \tab 3 \tab Bacteria\cr
  Crenarchaeota \tab 1 \tab 0 \tab 0 \tab Archaea\cr
  Ascomycota \tab 2 \tab 765 \tab 971 \tab Fungi 
}}

\item{melt.id}{A column name to \code{\link{melt}} and used to assign the colours.}

\item{title}{Graph title}

\item{x.lab, y.lab}{The label of x-axis or y-axis, such as plot names.}

\item{autoWidth}{If TRUE, then use number of bars and legend columns 
to estimate pdf width automatically. Default to TRUE.}

\item{...}{More from \code{\link{ggBarChart}}.}

\item{low, }{high Refer to \pkg{ggplot2} \code{\link{scale_fill_gradient}}. 
Default to low="white", high="steelblue".}
}
\description{
Percentage bar chart coloured by groups, which is extended from \code{\link{ggBarChart}}.
}
\examples{
data(reads.phyla)
reads.phyla
bar.chart <- ggPercentageBarChart(reads.phyla, melt.id="TaxaGroup")
bar.chart$gg.plot
pdf.ggplot(bar.chart$gg.plot, fig.path="taxa-percentage-bar.pdf", width=bar.chart$pdf.width, height=8) 
}
\keyword{graph}

