% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExtended.R
\name{gtNMDSPlot}
\alias{gtNMDSPlot}
\title{NMDS Plot}
\usage{
gtNMDSPlot(comm, attr.df, colour.id = NULL, shape.id = NULL,
  link.id = NULL, add.text = TRUE, text.data = NULL, text.size = 3,
  palette = NULL, xintercept = NULL, yintercept = NULL,
  distance = "bray", title = "MDS", verbose = TRUE, ...)
}
\arguments{
\item{comm}{Community data for NMDS plot. 
It is dissimilarities either as a \code{\link{dist}} structure 
or as a community data defined by \code{\link{vegan}} which is 
a transposed matrix of community matrix in \code{\link{ComMA}}.
Use \code{\link{transpose.df}} to rotate the data frame.
Detail to \code{\link{metaMDS}}.}

\item{attr.df}{A data frame to define how NMDS plot is coloured, shaped, or linked.
where its row names have to match row names of community data \code{comm}.}

\item{distance}{Dissimilarity index used in \code{\link{vegdist}}.}

\item{...}{Other arguments passed to \code{\link{gtScatterPlot}}.}
}
\description{
Use \code{\link{metaMDS}} in \code{\link{vegan}}
to create Nonmetric Multidimensional Scaling (NMDS) plot.
It is extended from \code{\link{gtScatterPlot}}.
}
\examples{

nmds.plot <- gtNMDSPlot(comm, env, colour.id="FishSpecies", shape.id="FeedingPattern", add.text=T)
plot(nmds.plot)
}
\keyword{graph}

