% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversities.R
\name{JostDiversity}
\alias{JostDiversity}
\alias{TurnoverDist}
\alias{calculateDissimilarityMatrix}
\alias{diversityTable}
\title{Jost diversity from \pkg{vegetarian} Package}
\usage{
diversityTable(t.community.matrix)

calculateDissimilarityMatrix(t.community.matrix, diss.fun = "beta1-1",
  printProgressBar)

TurnoverDist(t.community.matrix)
}
\arguments{
\item{t.community.matrix}{is abundances argument in \pkg{vegetarian} \code{\link{d}}, 
which is a transposed matrix of community matrix, 
where rows are plots (Use plots instead of subplots.), columns are OTUs.}

\item{diss.fun}{Similarity/dissimilarity index, values are jaccard, horn.morisita, 
bray.curtis, and beta1-1. Default to beta1-1, but it is slower than other indices.}

\item{printProgressBar}{Whether to print progress bar, if missing, it will be nrow(row.pairs)>100}
}
\value{
\code{diversityTable} returns a 3x3 data frame: columns are levels of diversity c("gamma", "alpha", "beta"), 
rows are orders of the diversity measure c(0, 1, 2). For example,
\tabular{rrrr}{
   \tab $q=0$ \tab $q=1$ \tab $q=1$\cr
  $D_\\gamma(q)$ \tab 13922.000000 \tab 2501.693162 \tab 601.509610\cr
  $D_\\alpha(q)$ \tab 2238.392857 \tab 880.944977 \tab 251.127187\cr
  $D_\\beta(q)$ \tab 6.219641 \tab 2.839784 \tab 2.395239 
}

\code{calculateDissimilarityMatrix} returns a \code{\link{matrix}} of similarity/dissimilarity, 
whose rows and columns are sample names

\code{TurnoverDist} returns a \code{\link{dist}} composed of pair-wise turnovers.
}
\description{
Data input \strong{t.community.matrix} is 
a transposed matrix of community matrix we defined in \pkg{ComMA}.
Community matrix from file is a matrix where rows are OTUs or individual species 
and columns are sites or samples. See \code{\link{ComMA}}.

Calculate similarity/dissimilarity distance matrix between samples.

Calculate pair-wise turnovers between samples.
}
\examples{
diversity.table <- diversityTable(t.community.matrix)

diss.matrix <- calculateDissimilarityMatrix(t.community.matrix, diss.fun="jaccard")

turnover.dist <- TurnoverDist(t.community.matrix)

}
\keyword{diversity}

