% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylRollingMedian.R
\name{methylRollingMedian}
\alias{methylRollingMedian}
\title{Title}
\usage{
methylRollingMedian(
  df,
  position_col,
  methyl_col,
  w_size,
  genome_size = 4641652,
  method = "exact"
)
}
\arguments{
\item{df}{A df with columns position_col and methyl_col.}

\item{position_col}{A string.}

\item{methyl_col}{A string.}

\item{w_size}{An integer for sliding window size.}

\item{genome_size}{An integer, by default the E. coli K-12 MG1655 genome size.}

\item{method}{A string "exact" (default) or "fast"}
}
\value{
A dataframe.
}
\description{
Title
}
\examples{
sliding_window_methylation <- methylRollingMedian(WT_average, position_col = 'Position', methyl_col = 'beta', w_size = 10000, method = "exact")
}
