% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\docType{class}
\name{MicrobeMethyl-class}
\alias{MicrobeMethyl-class}
\alias{MicrobeMethyl}
\title{MicrobeMethyl class}
\description{
An S4 class to represent adenine methylation (6mA) data from a single
bacterial sample, particularly in E. coli. This class stores methylation
data (beta values and number of reads), sample name, and sample metadata.
}
\section{Slots}{

\describe{
\item{\code{sample_name}}{A character string representing the name of the sample.}

\item{\code{metadata}}{A list containing metadata for the sample, such as
experimental conditions and any other relevant information.}

\item{\code{assembly}}{A character string with the chromosome name.}
}}

