% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeMethyl.R
\docType{class}
\name{microbeMethylExperiment-class}
\alias{microbeMethylExperiment-class}
\alias{microbeMethylExperiment}
\title{microbeMethylExperiment class}
\description{
An S4 class to represent an experiment containing multiple microbeMethyl
objects. This class stores a list of microbeMethyl objects and experiment
metadata.
}
\section{Slots}{

\describe{
\item{\code{samples}}{A list of microbeMethyl objects, each representing a single
sample from the same experiment.}

\item{\code{experiment_metadata}}{A list containing metadata for the experiment,
such as experimental conditions, date, and any other relevant information.}
}}

