% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPrioritisation.R
\name{PlotPrioritisation}
\alias{PlotPrioritisation}
\alias{getPlotPrior}
\alias{getPlotPrior.JostDiver}
\alias{getPlotPrior.PhyloAlpha}
\title{Plot prioritisation}
\usage{
getPlotPrior.JostDiver(t.community.matrix, lev = c("alpha", "beta", "gamma"),
  q = 1, order.by = c("sample", "rank", "diversity"))

getPlotPrior.PhyloAlpha(t.community.matrix, phylo.tree, taxa.match = TRUE,
  order.by = c("sample", "rank", "diversity"), ...)

getPlotPrior(cm.list, is.transposed = FALSE, tre.list = list(),
  diversities = c("gamma0", "gamma1", "beta0", "beta1", "pd.alpha",
  "sp.rich"))
}
\arguments{
\item{t.community.matrix}{A transposed matrix from community matrix, 
where rows are plots (Use plots instead of subplots.), columns are OTUs.}

\item{lev}{Level of diversity to be calculated. 
Will accept: 'alpha', 'beta', or 'gamma'.}

\item{q}{Order of the diversity measure. 
Defaults to the Shannon case where q = 1.}

\item{order.by}{How the result is ordered. 
Choose from 'sample', 'rank', or 'diversity'.}

\item{phylo.tree, ...}{The parameters passed to \code{\link{phylo.alpha}}.}

\item{taxa.match}{Logical, if taxa in phylogenies do not match OTUs in the community. 
If TRUE, as default, to use t.community.matrix and phylo.tree directly, 
otherwise to call \code{\link{match.phylo.comm}} 
before calculate phylogenetic alpha diversity.}

\item{cm.list}{The list of community matrices.}

\item{is.transposed}{If TRUE, then the community matrix is already
transposed to be the valid input of \code{\link{vegdist}}.  
Default to FASLE to transpose.}

\item{tre.list}{A list of phylo tree objects for 'pd.alpha' and 'sp.rich',
corresponding to \code{cm.list}. Default to an empty list.}

\item{diversities}{The vector of diversities used to compute plot prioritisation.
The values are 'gamma0','gamma1','beta0','beta1','pd.alpha','sp.rich'. 
The first two are calculated by \code{\link{d}}, 
the last two by \code{\link{pd}}.}
}
\value{
A data frame with 2 columns: \emph{rank, diversity}. 
Rank 1 is the most important plot, 
\emph{n} is the least important, and row.names are plot names. 
For example,
\tabular{rrr}{
   \tab rank \tab diversity\cr
  CM30c39 \tab 28 \tab 1845.785714\cr
  CM30c44 \tab 27 \tab 1875.888889\cr
  CM31a5 \tab 26 \tab 1899.653846 
}
}
\description{
Ranking of sample plots by their contributions 
to the total biodiversity.
}
\details{
\code{getPlotPrior.JostDiver} computes plot prioritisation 
by Jost diversity calculated from \pkg{vegetarian} \code{\link{d}}.
It uses a greedy algorithm to remove plots sequentially 
so as to minimize the loss of diversity among the remaining plots,
which always chooses the 1st plot if there are multi-results 
in each prioritisation loop.
Rank 1 is the most important plot and removed at the last, 
\emph{n} is the least important and removed in the beginning.

\code{getPlotPrior.PhyloAlpha} calculates plot prioritisation 
by phylogenetic alpha diversity from \code{\link{phylo.alpha}}.
It also can return the ranks based on species richness (SR), 
but they may be different to ranks calculated from 
\code{getPlotPrior.JostDiver} using gamma0 (also species richness).

\code{getPlotPrior} is a generic function including both 
\code{getPlotPrior.JostDiver} and \code{getPlotPrior.PhyloAlpha}, 
and it also handles multiple communities.
}
\examples{
plot.prior.g1 <- getPlotPrior.JostDiver(t.community.matrix, lev="gamma", q=1)

phylo.alpha <- getPlotPrior.PhyloAlpha(t.community.matrix, phylo.tree)

plot.prior <- getPlotPrior(cm.list, is.transposed=FALSE, tre.list=tre.list, diversities=c("gamma1","beta1","pd.alpha","sp.rich"))

}
\keyword{plot}
\keyword{prioritisation}

