% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedundancyAnalysis.R
\name{RDA}
\alias{RDA}
\alias{doRDA}
\alias{matchCMEnv}
\alias{plotCorrelations}
\alias{plotRDA}
\alias{printXTable.RDA}
\title{Redundancy Analysis}
\usage{
doRDA(tcm.or.dist, env, verbose = TRUE)

matchCMEnv(cm, env, is.transposed = FALSE, verbose = TRUE)

plotCorrelations(df.numeric, corr.gram = FALSE, cex.axis = 0.75,
  cex.cor = 0.9, col = "#333333")

plotRDA(rda, env, colour.id = "Elevation",
  title = "Backward RDA, Jaccard distance", x.lab = "", y.lab = "",
  palette = c("blue", "orange"), scale.limits.min = NULL,
  no.legend = FALSE, legend.title = "Elevation (m)", verbose = TRUE, ...)

printXTable.RDA(rda, matrix.name = "", taxa.group = "", table.file = NULL,
  invalid.char = FALSE)
}
\arguments{
\item{tcm.or.dist}{A transposed community matrix 
or dist object of distances between samples. 
Rows are samples.}

\item{env}{The enviornmental meta-data, where rows are samples, 
and they must be same as rownames(tcm.or.dist) inlcuding order.
In addition, make sure rownames (enviornmental variables) 
are valid to \code{\link{formula}}.}

\item{verbose}{More details. Default to TRUE.}

\item{cm}{A community matrix.}

\item{is.transposed}{If TRUE, then the community matrix is already
transposed to be the valid input of \code{\link{vegdist}}.  
Default to FASLE.}

\item{df.numeric}{The data frame or matrix containing 
numeric variables (columns) to plot.}

\item{corr.gram}{Logical, if use \code{\link{corrgram}} 
instead of \code{\link{plot}}.}

\item{rda}{The ordination result from \code{doRDA}.}

\item{colour.id}{The column name in \code{env} to colour points and texts.}

\item{title, x.lab, y.lab}{Title, x, y label.}

\item{palette}{Refer to \code{\link{ggOptPalette}} in \pkg{gg1L}. 
Default to c("blue", "orange").}

\item{scale.limits.min}{Manually set the minimum data range of the colour scale, 
for example, in legend. The code set \code{limits} in \code{\link{discrete_scale}}
to \code{c(scale.limits.min, max(df[,colour.id]))}.}

\item{no.legend, legend.title}{Configure legend}

\item{matrix.name}{The string to locate the matrix from its file name. 
Only used for table name and label here.}

\item{taxa.group}{The taxonomic group. Only used for table name and label here.}

\item{table.file}{If NULL, then print the results to console, 
otherwise print them to the file. Default to NULL.}

\item{rda}{The list of results from \code{doRDA}.}
}
\description{
Redundancy Analysis using \pkg{vegan} 
\code{\link{capscale}}.
}
\details{
\code{doRDA} makes Constrained Analysis 
of Principal Coordinates for 
eDNA data sets given environmental variables.

\code{matchCMEnv} matches the sample names between 
community matrix and the enviornmental meta-data including the order, 
in order to provide the valid input to RDA analysis \code{doRDA}.

Preprecessing can be applied by \code{\link{preprocessCM}} 
and \code{\link{preprocessEnv}}.

\code{plotCorrelations} plots numeric variables (columns).

Tip: use \code{\link{"\%<a-\%"}} in \pkg{pryr} to save plots.

\code{plotRDA} plots a ordination result from \code{doRDA}.

\code{printXTable.RDA} prints \code{\link{xtable}} given rda results.
}
\note{
Make sure you are using the correct data type 
for both the community matrix and enviornmental meta-data.
Run \code{sapply(env, class)} to check if the value 
should be discrete(character) or numeric. 
\code{\link{convertType}} will convert data frame columns 
to different type easily.
}
\examples{
# 1. get the community matrix and enviornmental meta-data
cm <- getCommunityMatrix("16S", min2=T, by.plot=F)
env <- getEnvData(by.plot=F)

# 2. preprocessing
cm.prep <- preprocessCM(cm, rm.samples=c("CM30b51","CM30b58"), min.abund=5, mean.abund.thr=0.025) 
env.prep <- preprocessEnv(env, rm.samples=c("CM30b51","CM30b58"), log.var=c(14:20), sel.env.var=c(4,5,8,9,14:22))
sapply(env, class)
 
# 3. match samples and transpose cm
tcm.env <- matchCMEnv(cm.prep, env.prep)

# 4. RDA
rda <- doRDA(tcm.env$tcm, tcm.env$env)

# 5. result
rda.pl <- plotRDA(rda[["backward"]], tcm.env$env)
rda.pl$plot
printXTable.RDA(rda, matrix.name="16S", taxa.group="all")



# already transposed
tcm.env <- matchCMEnv(tcm, env, is.transposed=T)
# Note colSums(cm) are based on samples
tcm.env <- matchCMEnv(cm, env)

# before RDA
require(pryr)
p \%<a-\% plotCorrelations(tcm.env$env)

rda.pl <- plotRDA(rda.list[[1]][["backward"]], env.prep, scale.limits.min=0)

printXTable.RDA(rda, matrix.name="16S", taxa.group="BACTERIA")

}
\keyword{rda}

