% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedundancyAnalysis.R
\name{RDA}
\alias{RDA}
\alias{matchCMEnv}
\alias{plotCorrelations}
\alias{printXTable.RDA}
\alias{proceedRDA}
\title{Redundancy Analysis}
\usage{
proceedRDA(tcm.or.dist, env, verbose = TRUE)

matchCMEnv(cm, env, is.transposed = FALSE, verbose = TRUE)

plotCorrelations(df.numeric, corr.gram = FALSE, cex.axis = 0.75,
  cex.cor = 0.9, col = "#333333")

printXTable.RDA(rda, table.file = NULL, invalid.char = FALSE,
  matrix.name = "", taxa.group = "")
}
\arguments{
\item{tcm.or.dist}{A transposed community matrix 
or dist object of distances between samples. 
Rows are samples.}

\item{env}{The enviornmental meta-data, where rows are samples, 
and they must be same as rownames(tcm.or.dist) inlcuding order.
In addition, make sure rownames (enviornmental variables) 
are valid to \code{\link{formula}}.}

\item{verbose}{More details. Default to TRUE.}

\item{cm}{A community matrix.}

\item{is.transposed}{If TRUE, then the community matrix is already
transposed to be the valid input of \code{\link{vegdist}}.  
Default to FASLE.}

\item{df.numeric}{The data frame or matrix containing 
numeric variables (columns) to plot.}

\item{corr.gram}{Logical, if use \code{\link{corrgram}} 
instead of \code{\link{plot}}.}

\item{rda}{The list of results from \code{proceedRDA}.}

\item{table.file}{If NULL, then print the results to console, 
otherwise print them to the file. Default to NULL.}

\item{matrix.name}{The string to locate the matrix from its file name. 
Only used for table name and label here.}

\item{taxa.group}{The taxonomic group. Only used for table name and label here.}
}
\description{
Redundancy Analysis using \pkg{vegan} 
\code{\link{capscale}}.
}
\details{
\code{proceedRDA} makes Constrained Analysis 
of Principal Coordinates for 
eDNA data sets given environmental variables.

\code{matchCMEnv} matches the sample names between 
community matrix and the enviornmental meta-data including the order, 
in order to provide the valid input to RDA analysis \code{proceedRDA}.

Preprecessing can be applied by \code{\link{preprocessCM}} 
and \code{\link{preprocessEnv}}.

\code{plotCorrelations} plots numeric variables (columns).

Tip: use \code{\link{"\%<a-\%"}} in \pkg{pryr} to save plots.

\code{printXTable.RDA} prints \code{\link{xtable}} given rda results.
}
\note{
Make sure you are using the correct data type 
for both the community matrix and enviornmental meta-data.
Run \code{sapply(env, class)} to check if the value 
should be discrete(character) or numeric. 
\code{\link{convertType}} will convert data frame columns 
to different type easily.
}
\examples{
# 1. get the community matrix and enviornmental meta-data
cm <- getCommunityMatrix("16S", min2=T, by.plot=F)
env <- getEnvData(by.plot=F)

# 2. preprocessing
cm.prep <- preprocessCM(cm, rm.samples=c("CM30b51","CM30b58"), min.abund=5, mean.abund.thr=0.025) 
env.prep <- preprocessEnv(env, rm.samples=c("CM30b51","CM30b58"), log.var=c(14:20), sel.env.var=c(4,5,8,9,14:22))
sapply(env, class)
 
# 3. match samples and transpose cm
tcm.env <- matchCMEnv(cm.prep, env.prep)

# 4. RDA
rda <- proceedRDA(tcm.env$tcm, tcm.env$env)

tcm.env <- matchCMEnv(tcm, env, is.transposed=T)
# Note colSums(cm) are based on samples
tcm.env <- matchCMEnv(cm, env)

# before RDA
require(pryr)
p \%<a-\% plotCorrelations(tcm.env$env)

printXTable.RDA(rda, table.file=NULL, matrix.name="16S", taxa.group="BACTERIA")

}
\keyword{rda}

