% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedundancyAnalysis.R
\name{RDA}
\alias{RDA}
\alias{plotCorrelations}
\alias{preprocessRDA}
\alias{printXTable.RDA}
\alias{proceedRDA}
\title{Redundancy Analysis}
\usage{
proceedRDA(tcm.or.dist, env, verbose = TRUE)

preprocessRDA(cm, env, is.transposed = FALSE, rm.samples = c(),
  min.abund = 0, sel.env.var = c(), log.var = c(), log.base = 2)

plotCorrelations(df.numeric, corr.gram = FALSE, cex.axis = 0.75,
  cex.cor = 0.9, col = "#333333")

printXTable.RDA(rda, table.file = NULL, invalid.char = FALSE,
  matrix.name = "", taxa.group = "")
}
\arguments{
\item{tcm.or.dist}{A transposed community matrix 
or dist object of distances between samples. 
Rows are samples.}

\item{env}{The enviornmental meta-data, where rows are samples, 
and they must be same as rownames(tcm.or.dist) inlcuding order.
In addition, make sure rownames (enviornmental variables) 
are valid to \code{\link{formula}}.}

\item{verbose}{More details. Default to TRUE.}

\item{cm}{A community matrix.}

\item{is.transposed}{If TRUE, then the community matrix is already
transposed to be the valid input of \code{\link{vegdist}}.  
Default to FASLE.}

\item{rm.samples}{Remove specified samples in a vector, 
it can be a keyword shared in sample names.
The vector will convert to a string separated by '|' to multi-samples. 
Default to empty vector to do nothing.}

\item{min.abund}{Exclude any samples with excessively 
low abundance. Defaul 0 to exaclude none.
The code is \code{tcm <- tcm[rowSums(tcm) > min.abund, ]}.}

\item{sel.env.var}{The vector of selected environmental variables, 
which can be colnames(env) or their indices. 
Defaul to an empty vector to choose all variables.}

\item{log.var, log.base}{It normally needs log transform to soil chemistry variables.
Use \code{\link{plotCorrelations}} to visualize variables and determine 
whether log transform should be applied. Default to no log transform.}

\item{df.numeric}{The data frame or matrix containing 
numeric variables (columns) to plot.}

\item{corr.gram}{Logical, if use \code{\link{corrgram}} 
instead of \code{\link{plot}}.}

\item{rda}{The list of results from \code{proceedRDA}.}

\item{table.file}{If NULL, then print the results to console, 
otherwise print them to the file. Default to NULL.}

\item{matrix.name}{The string to locate the matrix from its file name. 
Only used for table name and label here.}

\item{taxa.group}{The taxonomic group. Only used for table name and label here.}
}
\description{
Redundancy Analysis using \pkg{vegan} 
\code{\link{capscale}}.
}
\details{
\code{proceedRDA} makes Constrained Analysis 
of Principal Coordinates for 
eDNA data sets given environmental variables.

\code{preprocessRDA} subsets variables and samples to 
be included in analysis \code{proceedRDA}.
It is required that samples in community matrix must match 
the enviornmental meta-data, including the order. 
Teh function does not handle dist object of distances.

\code{plotCorrelations} plots numeric variables (columns).

Tip: use \code{\link{"\%<a-\%"}} in \pkg{pryr} to save plots.

\code{printXTable.RDA} prints \code{\link{xtable}} given rda results.
}
\examples{
rda <- proceedRDA(tcm.or.dist, env)

tcm.env <- preprocessRDA(tcm, env, is.transposed=T)
# Note colSums(cm) are based on samples
tcm.env <- preprocessRDA(cm, env, rm.samples=c("CM30b51","CM30b58"), min.abund=mean(colSums(cm))*0.025, sel.env.var=c(4,5,8,9,14:22))

# before RDA
require(pryr)
p \%<a-\% plotCorrelations(tcm.env$env)

printXTable.RDA(rda, table.file=NULL, matrix.name="16S", taxa.group="BACTERIA")

}
\keyword{rda}

