% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedundancyAnalysis.R
\name{RDA}
\alias{RDA}
\alias{preprocessRDA}
\alias{printXTable.RDA}
\alias{proceedRDA}
\title{Redundancy Analysis}
\usage{
proceedRDA(tcm.or.dist, env, verbose = TRUE)

preprocessRDA(tcm, env)

printXTable.RDA(rda, table.file = NULL, invalid.char = FALSE,
  matrix.name = "", taxa.group = "")
}
\arguments{
\item{tcm.or.dist}{A transposed community matrix 
or dist object of distances between samples. 
Rows are samples.}

\item{env}{The enviornmental meta-data, where rows are samples, 
and they must be same as rownames(tcm.or.dist) inlcuding order.}

\item{verbose}{More details. Default to TRUE.}

\item{tcm}{A transposed community matrix. 
Rows are samples.}

\item{rda}{The list of results from \code{proceedRDA}.}

\item{table.file}{If NULL, then print the results to console, 
otherwise print them to the file. Default to NULL.}

\item{matrix.name}{The string to locate the matrix from its file name. 
Only used for table name and label here.}

\item{taxa.group}{The taxonomic group. Only used for table name and label here.}
}
\description{
Redundancy Analysis using \pkg{vegan} 
\code{\link{capscale}}.
}
\details{
\code{proceedRDA} makes Constrained Analysis 
of Principal Coordinates for 
eDNA data sets given environmental variables.

\code{preprocessRDA} preprocesses data for \code{proceedRDA}.
It is required, if samples in community matrix are not matching 
the enviornmental meta-data, including the order. 
It does not handle dist object of distances.

\code{printXTable.RDA} prints \code{\link{xtable}} given rda results.
}
\examples{
rda <- proceedRDA(tcm.or.dist, env)

tcm.env <- preprocessRDA(tcm, env)

printXTable.RDA(rda, table.file=NULL, matrix.name="16S", taxa.group="BACTERIA")

}
\keyword{rda}

