% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readFileLineByLine}
\alias{readFileLineByLine}
\title{Read File Line By Line}
\usage{
readFileLineByLine(file)
}
\arguments{
\item{file}{The file to read.}
}
\value{
A character vector of length the number of lines read. Refer to \code{\link{readLines}}.
}
\description{
Read File Line By Line
}
\examples{
linn <- readFileLineByLine("time.txt")
}
\keyword{IO}

