% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialUtils.R
\name{spatial}
\alias{deg2rad}
\alias{geodesicDist.hf}
\alias{geodesicDist.slc}
\alias{spatial}
\title{Utils related to spatial distances}
\source{
\url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}
\usage{
deg2rad(deg)

geodesicDist.slc(lati.long, file = NA, sep = "\\t")

geodesicDist.hf(lati.long, file = NA, sep = "\\t")
}
\arguments{
\item{deg}{The degrees.}

\item{lati.long}{A 2-column data frame contains latitude/longitude coordinates.}

\item{file}{Either a character string naming a file or a connection open for writing. 
"" indicates output to the console. If NA, as default, then do nothing.}

\item{sep}{The field separator string.}
}
\description{
The functions related to spatial distances.
}
\details{
Converts degrees to radians

\code{geodesicDist.slc} calculates the geodesic distance (km) 
between two points specified by radian latitude/longitude 
using the Spherical Law of Cosines (slc).

\code{geodesicDist.hf} calculates the geodesic distance (km) 
between two points specified by radian latitude/longitude 
using the Haversine formula (hf).
}
\examples{
lat1 <- deg2rad(deg)

dists.btw <- geodesicDist.slc(env.byplot[, c("Latitude","Longitude")])

dists.btw <- geodesicDist.hf(env.byplot[, c("Latitude","Longitude")])

}
\keyword{spatial}

