% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{stringUtils}
\alias{countCharOccurrences}
\alias{findDuplicates}
\alias{freqUniqueValue}
\alias{gusbDF}
\alias{simpleCap}
\alias{stringUtils}
\alias{substrBetween}
\alias{substrLast}
\alias{substrLastSplit}
\alias{trimSpace}
\alias{trimStartEndSpace}
\title{Utils to deal with strings and characters}
\usage{
trimStartEndSpace(x)

trimSpace(x)

substrLast(x, n)

substrLastSplit(x, split = "/", ...)

substrBetween(x, l, r)

simpleCap(x)

countCharOccurrences(char, s)

freqUniqueValue(x)

findDuplicates(x)

gusbDF(pattern, replacement, df, stringsAsFactors = FALSE,
  check.names = FALSE, ...)
}
\arguments{
\item{x}{The string.}

\item{n}{The last n characters}

\item{split}{character vector (or object which can be coerced to such) containing regular expression(s) 
(unless fixed = TRUE) to use for splitting. Detail to \code{\link{strsplit}}.}

\item{...}{More parameters passed to \code{\link{strsplit}} from \code{substrLastSplit}.}

\item{l}{The regex on the left.}

\item{r}{The regex on the right.}
}
\description{
Useful function to manipulate strings and characters in \code{\link{ComMA}}.
}
\details{
\code{trimStartEndSpace} trims the leading and trailing whitespace in a string.
Refer to \url{http://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace-in-r}.

\code{trimSpace} removes all whitespace from a string.
Refer to \url{http://stackoverflow.com/questions/5992082/how-to-remove-all-whitespace-from-a-string}.

\code{substrLast} extracts the last n characters from a string x.
Refer to \url{http://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r}.

\code{substrLastSplit} extracts the last characters from a string 
\code{x} splitted by \code{split}. It can be used to get the file name 
from a file path, as default \code{split='/'}.

\code{substrBetween} extracts the substring between two given regular expressions.
Refer to \url{http://stackoverflow.com/questions/14146362/regex-extract-string-between}.

\code{simpleCap} capitalizes the first letter of a word string.
Refer to \url{http://stackoverflow.com/questions/6364783/capitalize-the-first-letter-of-both-words-in-a-two-word-string}.

\code{countCharOccurrences} returns the occurrences of given charater 
\code{char} in string \code{s}. Refere to 
\url{https://techoverflow.net/blog/2012/11/10/r-count-occurrences-of-character-in-string/}.

\code{freqUniqueValue} returns 2-column data frame of the unique values 
and their frequency given a vector \code{x}. Refere to 
\url{http://stackoverflow.com/questions/16905425/find-duplicate-values-in-r}.

\code{findDuplicates} uses \code{freqUniqueValue} to 
find the duplicates in a vector \code{x}.

\code{gusbDF} applies \code{\link{gsub}} to 
the entire data frame.
}
\examples{
text = "   foo bar  baz 3 "
trimStartEndSpace(text)
[1] "foo bar  baz 3"

text = "   foo bar  baz 3 "
trimSpace(text)
[1] "foobarbaz3"

x <- "some text in a string"
substrLast(x, 6)
[1] "string"
substrLast(x, 8)
[1] "a string"

filenames <- list.files(".", pattern="*.java", full.names=TRUE, recursive = TRUE)
substrLastSplit(filenames)
[1] "BEASTInterfaceTest.java"       "BooleanParameterListTest.java" "IntegerParameterListTest.java"

x <- "command took 0:1:34.67 (94.67s total)"
substrBetween(x, "\\\\(", "s total\\\\)")
[1] "string"

simpleCap("BACTERIA")
# for multi-words
taxaGroups <- c("BACTERIA", "FUNGI", "PROTISTS", "ANIMALIA")
sapply(taxaGroups, simpleCap)
#BACTERIA      FUNGI   PROTISTS   ANIMALIA 
#"Bacteria"    "Fungi" "Protists" "Animalia"

countCharOccurrences("a", "application")

freqUniqueValue(c("a", "b", "a"))
#    Var1 Freq
# 1    a    2
# 2    b    1
freqUniqueValue(c(1, 1, 2, 1))

findDuplicates(c("a", "b", "a"))

gusbDF(".00", "", df)

}
\keyword{utils}

