% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracer.R
\name{Tracer}
\alias{Tracer}
\alias{getTraces}
\alias{readMCMCLog}
\title{Tracer in R}
\usage{
readMCMCLog(file, ...)

getTraces(mcmc.log, burn.in = NULL, burn.in.perc = 0.1, verbose = TRUE)
}
\arguments{
\item{file}{The file to read/write.}

\item{...}{Other arguments passed to \code{\link{readFile}}.}

\item{mcmc.log}{The data frame from MCMC log file whose 
column names are parameters and row names are the number 
of states at each sample.}

\item{burn.in}{The number of states of burn in. Default to NULL 
to use \code{burn.in.perc}. Otherwise, use \code{burn.in}.}

\item{burn.in.perc}{The percentage of total chain length to be burn in. 
Default to 0.1. Only used if \code{burn.in} is NULL.}
}
\description{
A simplified implementation of Tracer \url{http://beast.bio.ed.ac.uk/tracer} in R.  
It can read MCMC log files from BEAST 1 \url{http://beast.bio.ed.ac.uk},
BEAST 2 \url{http://www.beast2.org}, or MrBayes \url{http://beast.bio.ed.ac.uk}.

It aims to make batch processing of multipe logs easier.
}
\details{
\code{readMCMCLog} reads MCMC log files and return a data frame 
whose column names are parameters and row names are the number 
of states at each sample.

\code{getTraces} preprocesses the data frame returned from 
\code{readMCMCLog}.
If \code{burn.in} is NULL, then use \code{burn.in.perc}, else 
consider \code{burn.in} first.
}
\examples{
mcmc.log <- readMCMCLog("data-raw/star.beast.log")

mcmc.traces <- getTraces(mcmc.log)

}
\keyword{Tracer}

