% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsCombine.R
\name{UtilsCombine}
\alias{UtilsCombine}
\alias{combineTriMatrix}
\alias{combineTwoDF}
\alias{getTriMatrix}
\alias{mergeByRownames}
\title{Utils to combine data frames or matrices into the required data format}
\usage{
getTriMatrix(df, row.col = c("l1", "l2"), value = "corr", na.to.0 = TRUE)

combineTriMatrix(tri.m1, tri.m2)

combineTwoDF(dfm, dfm2, rm.zero = TRUE, return.df = TRUE, ...)

mergeByRownames(x, y, warning.msg = TRUE, ...)
}
\arguments{
\item{df}{The data frame containing pairwised comparison result, 
such as correlations.  
For example, a data frame can be 
\tabular{rrrr}{
  l1 \tab l2 \tab corr\tab sign\cr
  16S \tab 18S \tab 0.827 \tab 0.001\cr
  16S \tab ITS \tab 0.585 \tab 0.001\cr
  18S \tab ITS \tab 0.729 \tab 0.001
}}

\item{value}{The column's values are going to fill in the output matrix.}

\item{na.to.0}{Logical, replace all NA to 0, as default.}

\item{tri.m1, tri.m2}{Two symmetric triangular matrices, 
which can be the output from \code{getTriMatrix}.}

\item{dfm}{A data frame or matrix.}

\item{dfm2}{The 2nd data frame or matrix whose values are into brackets.}

\item{rm.zero}{Default to TRUE to remove all " (0)".}

\item{return.df, ...}{Default to TRUE to return a data frame, otherwise a matrix.}

\item{...}{pass to \code{\link{merge}}.}

\item{x, y}{data frames, or objects to be coerced to one.}

\item{warning.msg}{logical; if TRUE, then print warning message when rows are missing after merge.}

\item{key}{Two column names containing the pair names,
which determine the output matrix's row names and column names.}
}
\description{
\code{combineTriMatrix} combines two symmetric triangular matrices into one.
The 1st matrix's upper triangle combines with the 2nd's lower triangle.

\code{combineTwoDF} combines two data frames or matrices with a same structure in one, 
put all values in 2nd data frame into brackets.

\code{mergeByRownames} merges two data frames by 'row.names' using \code{\link{merge}}.
}
\details{
\code{getTriMatrix} converts pairwised comparison result 
into a symmetric triangular matrix.
The pairwised comparison result is stored in a data frame.
Two of their columns must have pair names in the same order, 
which determine the matrix's row names and column names.
The 1st data frame goes to the lower triangle of the matrix, 
and the 2nd is in upper triangle.
}
\examples{
corr.tri <- getTriMatrix(corr.df)

corr.sign.tri <- combineTriMatrix(corr.tri, sign.tri)

df <- combineTwoDF(df, df2, stringsAsFactors=FALSE)

df <- mergeByRownames(df, df2)

}
\keyword{utils}

