% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsSequences.R
\name{utilsSeq}
\alias{subsetSequences}
\alias{utilsSeq}
\title{Utils to manipulate sequences or OTUs}
\usage{
subsetSequences(otus.file, otus.names, otus.out.file,
  regex1 = "(\\\\|[0-9]+)", regex2 = "", ignore.case = TRUE,
  max.seq = NA)
}
\arguments{
\item{otus.file}{The fasta file of OTU representive sequences. 
Read by \code{\link{readFasta}}.}

\item{otus.names}{The vector of names to match sequence labels in \code{otus.file}.}

\item{otus.out.file}{The output fasta file containing extrated sequences.}

\item{regex1, regex2}{Use for \code{\link{gsub}(regex1, regex2, row.names)} 
to remove or replace annotation from original labels. 
Default to \code{regex1="(\\|[0-9]+)", regex2=""}, 
which removes size annotation seperated by "|".}

\item{ignore.case}{Refer to \code{\link{gsub}}.}

\item{max.seq}{Give the number (\code{max.seq}) of seleceted sequences randomly in the end. 
Defaul to NA to ignore it.}
}
\description{
Utils to manipulate sequences or OTUs, 
such as extract OTUs given the subset of OTU names.
They are depended on Bioconductor packages  
\url{http://www.bioconductor.org/}.
}
\details{
\code{subsetSequences} returns the subset of OTUs matching given OTU names. 
It is depended on \code{\link{ShortRead}} package. Follow the instruction 
\url{https://bioconductor.org/packages/release/bioc/html/ShortRead.html} to install.
}
\examples{
subsetSequences(otus.file, otus.names, otus.out.file)

}
\keyword{util}

