% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsTaxa.R
\name{utilsTaxa}
\alias{assignTaxaByRank}
\alias{mergeCMTaxa}
\alias{subsetTaxaTable}
\alias{utilsTaxa}
\title{Utils to preprocess taxa table}
\usage{
subsetTaxaTable(taxa.table, taxa.group = "assigned", rank = "kingdom",
  include = TRUE, verbose = TRUE)

mergeCMTaxa(community.matrix, taxa.table, classifier = "MEGAN",
  min.conf = 0.8, has.total = 1, sort = TRUE, col.ranks = c("kingdom",
  "phylum", "class", "order", "family", "genus"))

assignTaxaByRank(cm.taxa, unclassified = 0)
}
\arguments{
\item{taxa.table}{A data frame to contain taxonomic classifications of OTUs. 
Columns are taxonomy at the rank or lineage, rows are OTUs which need to 
match rows from community matrix. Use \code{\link{readTaxaTable}} to get it from file.}

\item{taxa.group}{The taxonomic group, the values can be 'all', 'assigned', or 
Group 'all' includes everything.
Group 'assigned' removes all uncertain classifications including 
'root', 'cellular organisms', 'No hits', 'Not assigned'. 
Alternatively, any high-ranking taxonomy in your taxonomy file 
can be used as a group or multi-groups (seperated by "|"), 
such as 'BACTERIA', 'Proteobacteria', etc. But they have to be 
in the same rank column in the file. Default to remove all 
uncertain classifications, even when group(s) assigned.}

\item{rank}{The rank column in the file to specify where to 
search for \code{taxa.group}.}

\item{include}{Define whether include or exclude given \code{taxa.group}. 
Default to TRUE.}

\item{verbose}{More details. Default to TRUE.}

\item{community.matrix}{Community matrix (OTU table), where rows are 
OTUs or individual species and columns are sites or samples. See \code{\link{ComMA}}.}

\item{classifier}{The classifier is used to generate \code{taxa.table}. 
Value is MEGAN or RDP. Default to MEGAN.}

\item{min.conf}{The confidence threshold to drop rows < \emph{min.conf}.}

\item{has.total}{If 0, then only return abudence by samples (columns) of community matrix. 
If 1, then only return toal abudence. If 2, then return abudence by samples (columns) and total. 
Default to 1.}

\item{sort}{Sort the taxonomy rank by rank. Default to TRUE.}

\item{col.ranks}{A vector or string of column name(s) of taxonomic ranks in the taxa table, 
which will determine the aggregated abundence matrix. They have to be full set or subset of 
c("superkingdom", "kingdom", "phylum", "class", "order", "family", "genus", "species"). 
Default to c("kingdom", "phylum", "class", "order", "family", "genus").}

\item{cm.taxa}{The data frame combined community matrix with 
taxonomic classifications generated by \code{mergeCMTaxa}. It 
should have attributes \code{ncol.cm} and \code{col.ranks}.}

\item{unclassified}{An interger to instruct how to deal with 
"unclassified" taxa. Default to 0, which keeps all "unclassified"
but moves them to the last rows. 
If 1, then remove the row whose taxon name is exact "unclassified". 
See the detail. 
If 2, then remove the row whose taxon name is exact "unclassified", 
but also merge all the rest "unclassified ???" to "unclassified rank",
such as "unclassified family".
If 3, then remove every rows containing "unclassified".}
}
\value{
\code{ncol.cm} and \code{col.ranks} are attributes of \code{cm.taxa} generated by \code{mergeCMTaxa}.

\code{ncol.cm} indicates how many column(s) is/are abundence in \code{cm.taxa}.

\code{col.ranks} records what ranks column(s) is/are in \code{cm.taxa}, 
which is also the input of \code{mergeCMTaxa}.
}
\description{
Utils to preprocess taxa table, 
and make it easy for visualization.
}
\details{
\code{subsetTaxaTable} takes or excludes a subset of given a taxa table at given rank.

\code{mergeCMTaxa} creates a data frame \code{cm.taxa} combined community matrix with 
taxonomic classification table. The 1st column is "row.names" that are OTUs/individuals, 
the next "ncol.cm" columns are abundence that can be sample-based or total, 
and the last "length(col.ranks)" columns are the ranks. 

All sequences either classified as 
"root|cellular organisms|No hits|Not assigned|unclassified sequences"
from BLAST + MEGAN, or confidence < \emph{min.conf} threshold from RDP, 
are changed to "unclassified", which will be moved to the last row.

\code{assignTaxa} provides a list of taxonomic assignments with abudence 
from community matrix at different rank levels. The function is iterated 
through \code{col.ranks}, and \code{\link{aggregate}}s abudence into 
taxonomy based on the rank in \code{col.ranks}.
}
\examples{
tt.sub <- subsetTaxaTable(tt.megan, taxa.group="Proteobacteria", rank="phylum")
tt.sub <- subsetTaxaTable(tt.megan, taxa.group="Cnidaria|Brachiopoda|Echinodermata|Porifera", rank="phylum", include=FALSE)

cm.taxa <- mergeCMTaxa(community.matrix, tt.megan) 
ta.megan <- assignTaxaByRank(cm.taxa)

cm.taxa <- mergeCMTaxa(community.matrix, tt.rdp, col.ranks="kingdom", classifier="RDP")
ta.rdp <- assignTaxaByRank(cm.taxa)

}
\keyword{utils}

