% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsTaxa.R
\name{utilsTaxa}
\alias{assignTaxaByRank}
\alias{groupsTaxaMembers}
\alias{mergeCMTaxa}
\alias{subsetTaxaTable}
\alias{utilsTaxa}
\title{Utils to preprocess taxa table}
\usage{
subsetTaxaTable(taxa.table, taxa.group = "assigned", rank = "kingdom",
  include = TRUE, verbose = TRUE)

mergeCMTaxa(community.matrix, taxa.table, classifier = "MEGAN",
  min.conf = 0.8, has.total = 1, sort = TRUE, col.ranks = c("kingdom",
  "phylum", "class", "order", "family", "genus"))

assignTaxaByRank(cm.taxa, unclassified = 0, aggre.FUN = sum)

groupsTaxaMembers(taxa.assign, cm.taxa, rank = "phylum",
  rm.unclassified = TRUE, regex1 = "(\\\\|[0-9]+)", regex2 = "",
  ignore.case = TRUE)
}
\arguments{
\item{taxa.table}{A data frame to contain taxonomic classifications of OTUs. 
Columns are taxonomy at the rank or lineage, rows are OTUs which need to 
match rows from community matrix. Use \code{\link{readTaxaTable}} to get it from file.}

\item{taxa.group}{The taxonomic group, the values can be 'all', 'assigned', or 
Group 'all' includes everything.
Group 'assigned' removes all uncertain classifications including 
'root', 'cellular organisms', 'No hits', 'Not assigned'. 
Alternatively, any high-ranking taxonomy in your taxonomy file 
can be used as a group or multi-groups (seperated by "|"), 
such as 'BACTERIA', 'Proteobacteria', etc. But they have to be 
in the same rank column in the file. Default to remove all 
uncertain classifications, even when group(s) assigned.}

\item{rank}{The rank to specify which column name in \code{taxa.table} to search.}

\item{include}{Define whether include or exclude given \code{taxa.group}. 
Default to TRUE.}

\item{verbose}{More details. Default to TRUE.}

\item{community.matrix}{Community matrix (OTU table), where rows are 
OTUs or individual species and columns are sites or samples. See \code{\link{ComMA}}.}

\item{classifier}{The classifier is used to generate \code{taxa.table}. 
Value is MEGAN or RDP. Default to MEGAN.}

\item{min.conf}{The confidence threshold to drop rows < \emph{min.conf}.}

\item{has.total}{If 0, then only return abudence by samples (columns) of community matrix. 
If 1, then only return toal abudence. If 2, then return abudence by samples (columns) and total. 
Default to 1.}

\item{sort}{Sort the taxonomy rank by rank. Default to TRUE.}

\item{col.ranks}{A vector or string of column name(s) of taxonomic ranks in the taxa table, 
which will determine the aggregated abundence matrix. They have to be full set or subset of 
c("superkingdom", "kingdom", "phylum", "class", "order", "family", "genus", "species"). 
Default to c("kingdom", "phylum", "class", "order", "family", "genus").}

\item{cm.taxa}{The data frame combined community matrix with 
taxonomic classifications generated by \code{mergeCMTaxa} (\code{\link{merge}}).
The 1st column is "row.names", \code{ncol.samples} columns 
abundence, and \code{length(col.ranks)} columns taxonomy 
at different ranks. 
It should also have attributes \code{ncol.cm} and \code{col.ranks}.}

\item{unclassified}{An interger to instruct how to deal with 
"unclassified" taxa. Default to 0, which keeps all "unclassified"
but moves them to the last rows. 
If 1, then remove the row whose taxon name is exact "unclassified". 
See the detail. 
If 2, then remove the row whose taxon name is exact "unclassified", 
but also merge all the rest "unclassified ???" to "unclassified rank",
such as "unclassified family".
If 3, then remove every rows containing "unclassified".}

\item{aggre.FUN}{A function for \code{FUN} in \code{\link{aggregate}}. 
Default to \code{sum} to provide the reads abundance. 
Make \code{aggre.FUN=function(x) sum(x>0)} provide the OTU abundance.}

\item{taxa.assign}{The data frame of taxonomic assignments with abudence
at the \code{rank}, where rownames are taxonomy at that rank, 
and columns are the sample names (may include total). It can be 
one element of the list generated by \code{assignTaxaByRank}. 
See the detail.}

\item{rm.unclassified}{Drop all unclassified rows (OTUs). Default to TRUE.}

\item{regex1, regex2}{Use for \code{\link{gsub}(regex1, regex2, row.names)} 
to remove or replace annotation from original labels. 
Default to \code{regex1="(\\|[0-9]+)", regex2=""}, 
which removes size annotation seperated by "|".}
}
\value{
\code{ncol.cm} and \code{col.ranks} are attributes of \code{cm.taxa} generated by \code{mergeCMTaxa}.

\code{ncol.cm} indicates how many column(s) is/are abundence in \code{cm.taxa}.

\code{col.ranks} records what ranks column(s) is/are in \code{cm.taxa}, 
which is also the input of \code{mergeCMTaxa}.
}
\description{
Utils to preprocess taxa table, 
and make it easy for visualization.
}
\details{
\code{subsetTaxaTable} takes or excludes a subset of given a taxa table at given rank.

\code{mergeCMTaxa} creates a data frame \code{cm.taxa} combined community matrix with 
taxonomic classification table. The 1st column is "row.names" that are OTUs/individuals, 
the next "ncol.cm" columns are abundence that can be sample-based or total, 
and the last "length(col.ranks)" columns are the ranks. 

All sequences either classified as 
"root|cellular organisms|No hits|Not assigned|unclassified sequences"
from BLAST + MEGAN, or confidence < \emph{min.conf} threshold from RDP, 
are changed to "unclassified", which will be moved to the last row.

\code{assignTaxa} provides a list of taxonomic assignments with abudence 
from community matrix at different rank levels, where rownames are taxonomy 
at that rank, and columns are the sample names (may include total). 
The function is iterated through \code{col.ranks}, and \code{\link{aggregate}}s 
abudence into taxonomy based on the rank in \code{col.ranks}.

\code{groupsTaxaMembers} groups the members (rows, also OTUs) from 
\code{cm.taxa} for each taxa in \code{taxa.assign} at the \code{rank}, 
and returns a list of members (OTUs) grouped by taxonomy. 
Default to drop all unclassified members (OTUs). 

It is impossible to trace back members after \code{assignTaxaByRank},
so that this function only has one option except the default, 
which assign the rest of members (OTUs) not picked up from other taxa  
into "unclassified". The result relies on using the identical \code{cm.taxa} 
in both \code{assignTaxaByRank} and \code{groupsTaxaMembers}.
}
\examples{
tt.sub <- subsetTaxaTable(tt.megan, taxa.group="Proteobacteria", rank="phylum")
tt.sub <- subsetTaxaTable(tt.megan, taxa.group="Cnidaria|Brachiopoda|Echinodermata|Porifera", rank="phylum", include=FALSE)

cm.taxa <- mergeCMTaxa(community.matrix, tt.megan) 
ta.megan <- assignTaxaByRank(cm.taxa)

cm.taxa <- mergeCMTaxa(community.matrix, tt.rdp, classifier="RDP", has.total=0)
ta.rdp <- assignTaxaByRank(cm.taxa, unclassified=2)
colSums(ta.rdp[["phylum"]])

taxa.members <- groupsTaxaMembers(ta.rdp[["phylum"]], tt.rdp)
taxa.members <- groupsTaxaMembers(ta.rdp[["family"]], tt.rdp, rank="family")

}
\keyword{utils}

