% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{writeTable}
\alias{writeTable}
\title{Write the data fram (table) to a file.}
\usage{
writeTable(df, file, append = FALSE, quote = FALSE, sep = "\\t",
  eol = "\\n", na = "NA", dec = ".", row.names = TRUE,
  col.names = TRUE, verbose = TRUE, msg.file = "file",
  msg.col = "columns", msg.row = "rows")
}
\arguments{
\item{df}{A data frame.}

\item{file}{If the file extension is \emph{csv}, 
then use \code{\link{write.csv}}, 
otherwise use \code{\link{write.table}}.}

\item{...}{More parameters, see \code{\link{write.table}}.}
}
\description{
Write the data fram (table) to a file.
}
\examples{
writeTable(df, file.path)
}
\keyword{IO}

